% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_edits.R
\name{add_marker}
\alias{add_marker}
\title{Function description}
\usage{
add_marker(
  input.seq,
  mrk,
  input.mat = NULL,
  tol = 0.001,
  verbose = TRUE,
  thresh.LOD.ph.to.insert = 10,
  thresh.rf.to.add = NULL
)
}
\arguments{
\item{void}{internal function}
}
\description{
Function description
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
