% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_rf.R
\name{aggregate_matrix}
\alias{aggregate_matrix}
\title{Aggregate matrix cells (lower the resolution by a factor)}
\usage{
aggregate_matrix(M, fact)
}
\arguments{
\item{M}{A numeric matrix to be aggregated.}

\item{fact}{An integer factor by which to lower the resolution.}
}
\value{
A list containing the aggregated matrix \code{R} and a list \code{markers} with the names of the markers in each aggregated cell.
}
\description{
Aggregate matrix cells (lower the resolution by a factor)
}
\keyword{internal}
