% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alfa_f1}
\alias{alfa_f1}
\title{Autotetraploid alfalfa F1 dataset.}
\format{
This list is of class \code{mappoly2.data} and contains the following components:

\describe{
\item{ploidy.p1}{The ploidy level of parent P1 (I195) is 4.}
\item{ploidy.p2}{The ploidy level of parent P2 (J432) is 4.}
\item{n.ind}{The total number of individuals is 184.}
\item{n.mrk}{The total number of unique markers is 2795, filtered for redundancy.}
\item{ind.names}{A character vector of the individuals' names.}
\item{mrk.names}{A character vector of the markers' names.}
\item{name.p1}{The name of parent P1 is I195.}
\item{name.p2}{The name of parent P2 is J432.}
\item{dosage.p1}{A named integer vector containing the dosage in parent P1 for all \code{n.mrk} markers.}
\item{dosage.p2}{A named integer vector containing the dosage in parent P2 for all \code{n.mrk} markers.}
\item{chrom}{A named character vector indicating the chromosome each marker belongs to.}
\item{genome.pos}{A named numeric vector containing the physical position of the markers in the sequence.}
\item{ref}{A character vector of the reference allele.}
\item{alt}{A character vector of the alternate allele.}
\item{geno.dose}{A numeric matrix containing the dosage for each marker (rows) for each individual (columns).}
\item{redundant}{A data frame containing two character vectors, \code{kept} and \code{removed}, of markers filtered for redundancy.}
\item{QAQC.values}{A list containing quality assurance and quality control values with the following components:
\itemize{
\item A data frame with statistics for each marker, including \code{miss} (missing data rate), \code{chisq.pval} (chi-squared test p-value), and \code{read.depth} (read depth).
\item A data frame with statistics for each individual, including \code{miss} (missing data rate) and \code{full.sib} (indicator of non-belonging to the analyzed bi-parental cross, generated by \code{\link[mappoly]{filter_individuals}}).
}
}
}
}
\usage{
alfa_f1
}
\description{
This dataset comprises 184 offspring from a cross between two tetraploid
alfalfa parents, I195 and J432, which are resistant and susceptible to
Aphanomyces euteiches, respectively. The biparental population was genotyped
with the alfalfa DArTag panel described
in \href{https://www.doi.org/10.46265/genresj.EMOR6509}{Zhao et al., (2023)}
}
\keyword{datasets}
