% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alfalfa_bc}
\alias{alfalfa_bc}
\title{Autotetraploid Alfalfa BC Dataset}
\format{
A list of class \code{mappoly2.data} with the following components:
(Structure is identical to the potato dataset above, with specifics adapted.)
}
\usage{
alfalfa_bc
}
\description{
This dataset includes 93 offspring from a backcross between tetraploid alfalfa
parents I195 and F1.85.209 (a hybrid derived from I195 and J432). Genotyping was
performed using the alfalfa DArTag panel as described in
\href{https://www.doi.org/10.46265/genresj.EMOR6509}{Zhao et al. (2023)}.
}
\keyword{datasets}
