% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alfalfa_f1}
\alias{alfalfa_f1}
\title{Autotetraploid Alfalfa F1 Dataset}
\format{
A list of class \code{mappoly2.data} with the following components:
(Structure is identical to the potato dataset above, with specifics adapted.)
}
\usage{
alfalfa_f1
}
\description{
This dataset contains 184 offspring from a cross between two tetraploid
alfalfa parents, I195 (resistant to Aphanomyces euteiches) and J432 (susceptible).
The population was genotyped using the alfalfa DArTag panel described in
\href{https://www.doi.org/10.46265/genresj.EMOR6509}{Zhao et al. (2023)}.
}
\keyword{datasets}
