% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apple}
\alias{apple}
\title{Diploid Apple F1 Dataset – 4-Chromosome Subset}
\format{
A list of class \code{mappoly2.data} with the following
components.
\emph{(Structure is identical to the \code{alfalfa_f1} dataset, with these
apple-specific values.)}
}
\usage{
apple
}
\description{
This dataset contains SNP dosage calls for \strong{318 F₁ seedlings} obtained from the
diploid apple cross \emph{‘Jonathan’ × ‘Golden Delicious’}.
The 610 bi-allelic SNPs included here belong to \strong{four chromosomes
(Chr 2, 3, 8 and 9)} and were generated with restriction-site–associated DNA
sequencing (RAD-seq) as described in Sun \emph{et al.} (2015).
The subset is provided for tutorials and vignette examples in \strong{\pkg{mappoly2}}.
}
\details{
\subsection{Details}{
\itemize{
\item \strong{Parents:} ‘Jonathan’ (maternal) and ‘Golden Delicious’ (paternal)
\item \strong{Population size:} 318 individuals
\item \strong{Markers:} 610 SNPs
\itemize{
\item Chr 2 – 134; Chr 3 – 160; Chr 8 – 149; Chr 9 – 167
}
\item \strong{Ploidy:} Diploid (2 × = 34)
}
}

\subsection{Source}{

Sun R, Chang Y, Yang F, \emph{et al.} (2015)
\emph{A dense SNP genetic map constructed using restriction site-associated DNA
sequencing enables detection of QTLs controlling apple fruit quality.}
\strong{BMC Genomics 16}:747. \url{doi:10.1186/s12864-015-1946-x}
}
}
\keyword{datasets}
