% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{augment_phased_map}
\alias{augment_phased_map}
\title{Phasing remaining markers based on pairwise recombination fraction estimation and multilocus estimation}
\usage{
augment_phased_map(
  input.seq,
  input.twopt,
  thresh.LOD.ph = 3,
  thresh.LOD.rf = 3,
  thresh.rf = 0.5,
  max.phases = 3,
  thresh.LOD.ph.to.insert = 10,
  thresh.rf.to.insert = NULL,
  tol = 0.001,
  verbose = TRUE
)
}
\arguments{
\item{void}{internal function}
}
\description{
Phasing remaining markers based on pairwise recombination fraction estimation and multilocus estimation
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
