% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{augment_phased_map_per_group}
\alias{augment_phased_map_per_group}
\title{Efficiently phase unprocessed markers in a given phased
genetic map, circumventing the need for complete
HMM-based map re computation.}
\usage{
augment_phased_map_per_group(
  x,
  gr,
  type = c("mds", "genome"),
  thresh.LOD.ph = 5,
  thresh.LOD.rf = 5,
  thresh.rf = 0.5,
  max.phase = 5,
  thresh.LOD.ph.to.insert = 10,
  thresh.rf.to.insert = NULL,
  tol = 0.001,
  verbose = TRUE
)
}
\arguments{
\item{void}{internal function}
}
\description{
Efficiently phase unprocessed markers in a given phased
genetic map, circumventing the need for complete
HMM-based map re computation.
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
