% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{calc_consensus_haplo}
\alias{calc_consensus_haplo}
\title{Calculate Haplotype Probabilities for Consensus Maps}
\usage{
calc_consensus_haplo(x, ncpus = 1)
}
\arguments{
\item{x}{An object of class 'mappoly2.consensus.map', representing consensus map data
for interconnected F1 populations.}

\item{ncpus}{The number of CPU cores to use for parallel processing, defaults to 1.
If more than 1 core is specified, parallel processing is used.}
}
\value{
The input object \code{x} with haplotype probabilities calculated for each consensus map.
}
\description{
This function calculates haplotype probabilities for each linkage group in a consensus
map dataset of interconnected F1 populations. It utilizes parallel processing for
efficient computation when multiple CPU cores are available.
}
\details{
The function processes the consensus map data to calculate haplotype probabilities
using a biallelic model. It leverages parallel processing capabilities when
multiple cores are available, significantly improving efficiency on large datasets.
}
