% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoprob.R
\name{calc_genoprob}
\alias{calc_genoprob}
\title{Conditional probabilities computation using Hidden Markov Models}
\usage{
calc_genoprob(
  input.seq,
  phase.conf = "all",
  rf = NULL,
  error = 0,
  verbose = FALSE,
  tol = 0.001
)
}
\arguments{
\item{void}{internal function}
}
\description{
Conditional probabilities computation using Hidden Markov Models
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
\keyword{internal}
