% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotypes.R
\name{calc_haplotypes_per_group}
\alias{calc_haplotypes_per_group}
\title{Conditional probabilities computation using Hidden Markov Models}
\usage{
calc_haplotypes_per_group(x, gr, type = c("both", "mds", "genome"))
}
\arguments{
\item{void}{internal function}
}
\description{
Conditional probabilities computation using Hidden Markov Models
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
\keyword{internal}
