% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_mrk_ind_from_QAQC}
\alias{.get_mrk_ind_from_QAQC}
\title{Retrieve Markers and Individuals Based on QA/QC Thresholds}
\usage{
.get_mrk_ind_from_QAQC(
  x,
  miss.mrk.thresh = +Inf,
  miss.ind.thresh = +Inf,
  chisq.pval.thresh = -Inf,
  read.depth.thresh = c(0, +Inf)
)
}
\arguments{
\item{x}{A list containing QA/QC values for markers and individuals, typically generated by \code{\link{.setQAQC}}.}

\item{miss.mrk.thresh}{A numeric value specifying the threshold for the missing data rate in markers (default is +Inf).}

\item{miss.ind.thresh}{A numeric value specifying the threshold for the missing data rate in individuals (default is +Inf).}

\item{chisq.pval.thresh}{A numeric value specifying the threshold for the chi-squared test p-value in markers (default is -Inf).}

\item{read.depth.thresh}{A numeric vector of length two specifying the lower and upper bounds for read depth in markers (default is c(0, +Inf)).}
}
\value{
A list with three elements:
\itemize{
\item \code{thresholds}: A list of the thresholds used for selection.
\item \code{mrk.names}: The names of the markers that meet the specified criteria.
\item \code{ind.names}: The names of the individuals that meet the specified criteria.
}
}
\description{
This internal function selects markers and individuals based on given QA/QC thresholds. It returns a list containing the names of markers and individuals that meet the specified criteria.
}
\keyword{internal}
