% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.map_skeleton}
\alias{.map_skeleton}
\title{Create a Map Skeleton for mappoly2 Data}
\usage{
.map_skeleton(x, mrk.id.list)
}
\arguments{
\item{x}{A mappoly2 data object; the function checks if the input is a valid mappoly2 data structure.}

\item{mrk.id.list}{A list or a vector of marker IDs for which the map skeleton is to be created.
If a vector is provided, it is converted into a list.}
}
\value{
Returns a mappoly2 sequence object with initialized map structures for each linkage group.
Each linkage group will have a map for MDS, genome, and work, but with \code{NULL} values for \code{order} and \code{phase}.
}
\description{
This function creates a skeleton structure for mapping data, specifically for mappoly2 data objects.
It initializes map structures for multiple linkage groups based on provided marker IDs.
}
\note{
This function is internal and usually not called directly by the user.
}
