% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.setQAQC}
\alias{.setQAQC}
\title{Set Quality Assurance and Quality Control (QA/QC) Values}
\usage{
.setQAQC(
  id.mrk,
  id.ind,
  miss.mrk = NA,
  miss.ind = rep(NA, length(id.ind)),
  read.depth = NA,
  full.sib = NA,
  chisq.pval
)
}
\arguments{
\item{id.mrk}{A vector of marker identifiers.}

\item{id.ind}{A vector of individual identifiers.}

\item{miss.mrk}{A numeric vector representing the missing data rate for each marker (default is NA).}

\item{miss.ind}{A numeric vector representing the missing data rate for each individual (default is a vector of NA of length equal to \code{id.ind}).}

\item{chisq.pval}{A numeric vector of chi-squared test p-values for each marker.}
}
\value{
A list with two elements:
\itemize{
\item \code{markers}: A data frame with the missing data rate (\code{miss}), chi-squared test p-values (\code{chisq.pval}), and read depth (set to \code{NA}) for each marker.
\item \code{individuals}: A data frame with the missing data rate (\code{miss}) and full sibling status (\code{full.sib}, set to \code{NA}) for each individual.
}
}
\description{
This internal function creates a list of QA/QC values for genetic markers and individuals.
It constructs two data frames: one for markers and another for individuals, each containing relevant QA/QC metrics.
}
\keyword{internal}
