% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_edits.R
\name{drop_marker}
\alias{drop_marker}
\title{Drop marker(s) from a sequence}
\usage{
drop_marker(input.seq, mrk, reestimate.map = TRUE)
}
\arguments{
\item{void}{internal function}
}
\description{
Drop marker(s) from a sequence
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
\keyword{internal}
