% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{edit_order}
\alias{edit_order}
\title{Edit sequence ordered by reference genome positions
comparing to another set order}
\usage{
edit_order(input.seq, group = 1, invert = NULL, remove = NULL)
}
\arguments{
\item{input.seq}{object of class mappoly2.sequence with alternative order (not genomic order)}

\item{group}{linkage group id}

\item{invert}{vector of marker names to be inverted}

\item{remove}{vector of marker names to be removed}
}
\description{
Edit sequence ordered by reference genome positions
comparing to another set order
}
