% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{embedded_to_numeric}
\alias{embedded_to_numeric}
\title{Removes all non-digit characters and then converts
the remaining string to an integer.}
\usage{
embedded_to_numeric(x)
}
\arguments{
\item{void}{internal function}
}
\description{
Removes all non-digit characters and then converts
the remaining string to an integer.
}
\keyword{internal}
