% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{estimate_consensus_map}
\alias{estimate_consensus_map}
\title{Estimate Consensus Genetic Map}
\usage{
estimate_consensus_map(
  x,
  err = 0,
  ncpus = 1,
  verbose = TRUE,
  detailed_verbose = FALSE,
  tol = 0.001,
  ret_H0 = FALSE
)
}
\arguments{
\item{x}{A list of 'mappoly2.prepared.integrated.data' objects.}

\item{err}{Error rate to be used in the HMM map estimation (default is 0.0).}

\item{ncpus}{Number of CPUs to use for parallel processing (default is 1).
Automatically adjusted to not exceed the number of available cores.}

\item{verbose}{Logical; if TRUE, enables the printing of progress messages (default is TRUE).}

\item{detailed_verbose}{Logical; if TRUE, enables the printing of detailed progress messages (default is FALSE).}

\item{tol}{Tolerance level for the estimation convergence (default is 10e-4).}

\item{ret_H0}{Logical; if TRUE, returns null hypothesis estimates (default is FALSE).}
}
\value{
A list of results from the consensus map estimation.
}
\description{
This function estimates a consensus genetic map among different biparental
populations.
}
