% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_data_to_csv}
\alias{export_data_to_csv}
\title{Export a \code{mappoly2.data} Object to a CSV File}
\usage{
export_data_to_csv(x, file = "file.csv", name.p1 = NULL, name.p2 = NULL)
}
\arguments{
\item{x}{An object of class \code{mappoly2.data}.}

\item{file}{A character string naming the output CSV file.}

\item{name.p1}{Optional name to replace the default label \code{P1} (parent 1).}

\item{name.p2}{Optional name to replace the default label \code{P2} (parent 2).}
}
\value{
Invisibly returns a \code{data.frame} containing the exported data.
}
\description{
This function exports a \code{mappoly2.data} object to a CSV file containing
SNP metadata and dosage information for both parents and all offspring.
}
