% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_data}
\alias{filter_data}
\title{Filter Genetic Data Based on Quality Metrics}
\usage{
filter_data(
  x,
  mrk.thresh = 0.1,
  ind.thresh = 0.1,
  chisq.pval.thresh = NULL,
  read.depth.thresh = c(5, 1000),
  plot.screening = TRUE
)
}
\arguments{
\item{x}{A \code{mappoly2.data} object containing genetic data.}

\item{mrk.thresh}{A numeric threshold for the missing data rate in markers (default is 0.10).}

\item{ind.thresh}{A numeric threshold for the missing data rate in individuals (default is 0.10).}

\item{chisq.pval.thresh}{A numeric threshold for chi-squared test p-values in markers (default is NULL, which sets the threshold using a Bonferroni approximation).}

\item{read.depth.thresh}{A numeric vector with two values indicating the lower and upper bounds for acceptable read depths in markers (default is c(5, 1000)).}

\item{plot.screening}{Logical, if TRUE (default), plots are generated to visually represent the screening process.}
}
\value{
Returns the input \code{mappoly2.data} object with additional components:
\itemize{
\item \code{screened.data}: A list containing the thresholds used for selection and the names of markers and individuals that met the specified criteria.
\item Class attribute \code{screened} is also appended to the object.
}
}
\description{
This function filters genetic data in a \code{mappoly2.data} object based on various quality control metrics. It applies thresholds for missing data rates, chi-squared p-values, and read depth to markers and individuals, and optionally plots the screening process.
}
\details{
The function first validates the input object, then applies the specified thresholds to filter out markers and individuals based on missing data rates, chi-squared p-values, and read depths. It updates the object with the results of this filtering and optionally generates plots to visualize the data before and after filtering.
}
\examples{
  filtered_data <- filter_data(B2721)

}
