% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_individuals_by_missing_rate}
\alias{filter_individuals_by_missing_rate}
\title{Filter Individuals Based on Missing Data Rate}
\usage{
filter_individuals_by_missing_rate(x, ind.thresh = 0.1, plot = TRUE)
}
\arguments{
\item{x}{A \code{mappoly2.data} object containing genetic data.}

\item{ind.thresh}{A numeric threshold for the missing data rate in individuals (default is 0.10).}

\item{plot}{Logical, if TRUE (default), generates a plot showing the frequency of missing data per individual.}
}
\value{
Returns the input \code{mappoly2.data} object with individuals not meeting the missing data threshold removed.
}
\description{
Filters individuals in a \code{mappoly2.data} object based on the missing data rate per individual.
Optionally, generates a plot showing the frequency of missing data for individuals and indicates which individuals were filtered out.
}
\details{
The function calculates the missing data rate for each individual and filters out individuals that exceed the specified threshold. If \code{plot = TRUE}, it generates a plot of missing data rates per individual, indicating which individuals were kept or filtered out.
}
\examples{
filtered_data <- filter_individuals_by_missing_rate(potato_dataset, ind.thresh = 0.05, plot = TRUE)

}
