% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_markers_by_missing_rate}
\alias{filter_markers_by_missing_rate}
\title{Filter Markers Based on Missing Data Rate}
\usage{
filter_markers_by_missing_rate(x, mrk.thresh = 0.1, plot = TRUE)
}
\arguments{
\item{x}{A \code{mappoly2.data} object containing genetic data.}

\item{mrk.thresh}{A numeric threshold for the missing data rate in markers (default is 0.10).}

\item{plot}{Logical, if TRUE (default), generates a plot showing the frequency of missing data per marker.}
}
\value{
Returns the input \code{mappoly2.data} object with markers not meeting the missing data threshold removed.
}
\description{
Filters markers in a \code{mappoly2.data} object based on the missing data rate per marker.
Optionally, generates a plot showing the frequency of missing data for markers and indicates which markers were filtered out.
}
\details{
The function calculates the missing data rate for each marker and filters out markers that exceed the specified threshold. If \code{plot = TRUE}, it generates a plot of missing data rates per marker, indicating which markers were kept or filtered out.
}
\examples{
filtered_data <- filter_markers_by_missing_rate(potato_dataset, mrk.thresh = 0.05, plot = TRUE)

}
