% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_markers_by_read_depth}
\alias{filter_markers_by_read_depth}
\title{Filter Markers Based on Read Depth}
\usage{
filter_markers_by_read_depth(x, read.depth.thresh = c(5, 1000), plot = TRUE)
}
\arguments{
\item{x}{A \code{mappoly2.data} object containing genetic data.}

\item{read.depth.thresh}{A numeric vector with two values indicating the lower and upper bounds for acceptable read depths in markers (default is c(5, 1000)).}

\item{plot}{Logical, if TRUE (default), generates a histogram of read depths per marker.}
}
\value{
Returns the input \code{mappoly2.data} object with markers not meeting the read depth thresholds removed.
}
\description{
Filters markers in a \code{mappoly2.data} object based on read depth thresholds.
Optionally, generates a histogram of read depths, indicating which markers were filtered out.
}
\details{
The function filters out markers with read depths outside the specified range. If \code{plot = TRUE}, it generates a histogram of read depths, indicating which markers were kept or filtered out.
}
\examples{
filtered_data <- filter_markers_by_read_depth(genetic_data, read.depth.thresh = c(10, 500), plot = TRUE)

}
