% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_rf_matrix}
\alias{filter_rf_matrix}
\title{Filter Recombination Fraction Matrix}
\usage{
filter_rf_matrix(
  x,
  type = c("rf", "sh"),
  thresh.LOD.ph = 0,
  thresh.LOD.rf = 0,
  thresh.rf = 0.5,
  mrk.names = NULL
)
}
\arguments{
\item{x}{A data object containing pairwise recombination fractions.}

\item{type}{Type of matrix to filter ("rf" for recombination fraction, "sh" for SH values).}

\item{thresh.LOD.ph}{LOD threshold for linkage phase (default = 0).}

\item{thresh.LOD.rf}{LOD threshold for recombination fraction (default = 0).}

\item{thresh.rf}{Recombination fraction threshold (default = 0.5).}

\item{mrk.names}{Marker names to include in the filtering (default = NULL, all markers).}
}
\value{
A filtered recombination fraction matrix or a list of filtered SH matrices.
}
\description{
This function filters a recombination fraction matrix based on LOD thresholds and recombination fraction limits.
}
