% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_segregation}
\alias{filter_segregation}
\title{Filter Markers Based on Chi-Square Test}
\usage{
filter_segregation(input.obj, chisq.pval.thres = NULL, inter = TRUE)
}
\arguments{
\item{input.obj}{The name of the input object of class \code{mappoly.data}.}

\item{chisq.pval.thres}{The p-value threshold used for the chi-square tests. By default,
this is the Bonferroni approximation with a global alpha of 0.05, i.e., 0.05/n.mrk.}

\item{inter}{If set to TRUE (default), this function will plot the distorted and
non-distorted markers.}
}
\value{
An object of class \code{mappoly.chitest.seq} that includes the following components:
\item{keep}{markers that follow Mendelian segregation pattern}
\item{exclude}{markers with distorted segregation}
\item{chisq.pval.thres}{threshold p-value used for chi-square tests}
\item{data.name}{name of the input dataset used to perform the chi-square tests}
}
\description{
This function filters markers based on the p-values obtained from a chi-square test.
The chi-square test assumes that the markers follow the expected segregation patterns under
Mendelian inheritance, random chromosome bivalent pairing, and no double reduction.
}
\author{
Marcelo Mollinari
\email{mmollin@ncsu.edu}
}
