% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{genome_order}
\alias{genome_order}
\title{Get the Genomic Position of Markers in a Sequence}
\usage{
genome_order(x, mrk.names, verbose = TRUE)
}
\arguments{
\item{x}{An object of class \code{mappoly2.sequence}.}

\item{mrk.names}{A vector of marker names for which the genomic position is required.}

\item{verbose}{Logical; if TRUE, progress messages will be printed.}
}
\value{
Returns a data frame with the genomic position of the specified markers,
ordered by chromosome and sequence position.
}
\description{
This internal function retrieves and orders the genomic position of markers in a given genetic sequence. It is primarily used within a larger analytical context.
}
\details{
The function checks for the availability of genomic position information
in the provided data object. If available, it orders the markers based
on their chromosome and sequence position. The function handles cases
where only chromosome information or only sequence position information
is available.
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
