% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dosage_type}
\alias{get_dosage_type}
\title{Determine Dosage Type of Markers}
\usage{
get_dosage_type(x, mrk.names)
}
\arguments{
\item{x}{An object containing dosage data.}

\item{mrk.names}{Names of the markers to be analyzed.}
}
\value{
A list categorizing each marker as simplex in parent 1, simplex in parent 2,
double simplex, or multiplex based on their dosage values.
}
\description{
Classifies each marker based on its dosage type using the provided dosage data.
}
