% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{get_genome_order}
\alias{get_genome_order}
\alias{print.mappoly2.geno.ord}
\alias{plot.mappoly2.geno.ord}
\title{Get the genomic position of markers in a sequence}
\usage{
get_genome_order(x, verbose = TRUE)

\method{print}{mappoly2.geno.ord}(x, ...)

\method{plot}{mappoly2.geno.ord}(x, ...)
}
\arguments{
\item{x}{an object of the class mappoly.geno.ord}

\item{verbose}{if \code{TRUE} (default), the current progress is shown; if
\code{FALSE}, no output is produced}

\item{...}{currently ignored}

\item{input.seq}{a sequence object of class \code{mappoly.sequence}}
}
\description{
This functions gets the genomic position of markers in a sequence and
return an ordered data frame with the name and position of each marker
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
