% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_markers_from_ordered_sequence}
\alias{get_markers_from_ordered_sequence}
\title{Extract Markers from Ordered Sequence}
\usage{
get_markers_from_ordered_sequence(
  x,
  lg,
  type = c("mds", "genome", "custom"),
  parent = c("p1p2", "p1", "p2")
)
}
\arguments{
\item{x}{An object containing marker maps.}

\item{lg}{Either a character vector naming the maps or a numeric vector indicating map indices.}

\item{type}{The type of sequence to consider (either "mds", "genome", or "custom").}
}
\value{
Markers based on the specified type.
}
\description{
Extracts markers from an ordered sequence in an object, based on the specified type.
}
