% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_markers_from_phased_sequence}
\alias{get_markers_from_phased_sequence}
\title{Extract Markers from Mapped Sequence}
\usage{
get_markers_from_phased_sequence(x, lg, type = c("mds", "genome", "custom"))
}
\arguments{
\item{x}{An object containing marker maps.}

\item{lg}{Either a character vector naming the maps or a numeric vector indicating map indices.}

\item{type}{The type of sequence to consider (either "mds", "genome", or "work").}
}
\value{
Markers based on the specified type for mapped sequences.
}
\description{
Similar to \code{get_markers_from_ordered_sequence} but specifically for mapped sequences.
}
