% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mrk_indices_from_chrom}
\alias{get_mrk_indices_from_chrom}
\title{Given a dataset containing chromosome information,
and a chromosome vector return a vector of screened
markers corresponding to the chromosomes provided}
\usage{
get_mrk_indices_from_chrom(x, chrom)
}
\description{
Given a dataset containing chromosome information,
and a chromosome vector return a vector of screened
markers corresponding to the chromosomes provided
}
