% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_QAQCmrk_indices}
\alias{get_QAQCmrk_indices}
\title{Given a screened data set, returns the indices of the
screened markers in the corresponding raw dataset}
\usage{
get_QAQCmrk_indices(x)
}
\description{
Given a screened data set, returns the indices of the
screened markers in the corresponding raw dataset
}
