% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sequence_metadata}
\alias{get_sequence_metadata}
\title{Given a dataset x and a vector containing the name of markers
w, returns metadata associated to w}
\usage{
get_sequence_metadata(x, w)
}
\description{
Given a dataset x and a vector containing the name of markers
w, returns metadata associated to w
}
