% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group}
\alias{group}
\title{Assign Markers to Linkage Groups}
\usage{
group(
  x = NULL,
  expected.groups = NULL,
  inter = TRUE,
  comp.mat = FALSE,
  LODweight = FALSE
)
}
\arguments{
\item{x}{A \code{mappoly2.data} object that contains recombination fraction information.}

\item{expected.groups}{The expected number of linkage groups (e.g., chromosomes)
for the species, if known. This parameter is mandatory unless \code{inter} is set to \code{TRUE}.}

\item{inter}{Logical; if \code{TRUE}, the function allows interactive determination of
the expected number of groups by plotting a dendrogram (default is \code{TRUE}).}

\item{comp.mat}{Logical; if \code{TRUE}, displays a comparison matrix between the
reference-based and linkage-based groupings if chromosome information is available (default is \code{FALSE}).}

\item{LODweight}{Logical; if \code{TRUE}, clustering is weighted by
the square of the LOD score (default is \code{FALSE}).}
}
\value{
An object of class \code{mappoly2.group}, which includes:
\describe{
\item{hc.snp}{A hierarchical clustering object generated by UPGMA or average method.}
\item{expected.groups}{The specified or interactively determined number of expected linkage groups.}
\item{groups.snp}{The assigned linkage groups for each marker.}
\item{seq.vs.grouped.snp}{Comparison matrix between genomic group information and the assigned groups, if \code{comp.mat} is \code{TRUE}.}
}
}
\description{
This function identifies linkage groups of markers using the results
from two-point (pairwise) analysis in a \code{mappoly2.data} or \code{mappoly2.sequence} object.
}
\details{
The function first validates the input object and then performs hierarchical clustering on the recombination fraction matrix.
If \code{LODweight} is \code{TRUE}, the clustering is weighted by the square of the LOD scores.
The function can interactively determine the number of expected groups and allows users to confirm or adjust this number.
A dendrogram is plotted for visual inspection. The function also optionally compares the assigned groups with genomic group information.
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
