% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{hmm_map_reconstruction}
\alias{hmm_map_reconstruction}
\title{Multipoint analysis using Hidden Markov Models}
\usage{
hmm_map_reconstruction(x, p, rf = NULL, verbose = FALSE, tol = 0.001)
}
\arguments{
\item{void}{internal function}
}
\description{
Multipoint analysis using Hidden Markov Models
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
\keyword{internal}
