% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{init_rf_filter}
\alias{init_rf_filter}
\title{Initialize Recombination Fraction Filtering}
\usage{
init_rf_filter(
  x,
  thresh.LOD.ph = 5,
  thresh.LOD.rf = 5,
  thresh.rf = 0.15,
  probs = c(0.05, 1),
  diag.markers = NULL,
  mrk.order = NULL,
  diagnostic.plot = TRUE,
  breaks = 100
)
}
\arguments{
\item{x}{An object of class \code{pairwise.rf}.}

\item{thresh.LOD.ph}{LOD threshold for linkage phase (default = 5).}

\item{thresh.LOD.rf}{LOD threshold for recombination fraction (default = 5).}

\item{thresh.rf}{Recombination fraction threshold (default = 0.15).}

\item{probs}{Probability range for filtering (default = c(0.05, 1)).}

\item{diag.markers}{Window of marker pairs to consider (default = NULL).}

\item{mrk.order}{Order of markers (used if diag.markers is not NULL).}

\item{diagnostic.plot}{Boolean to control the generation of a diagnostic plot (default = TRUE).}

\item{breaks}{Number of breaks for histogram in diagnostic plot (default = 100).}
}
\value{
A modified \code{pairwise.rf} object with filtered data.
}
\description{
Internal function to initialize filtering based on recombination fractions, LOD scores for linkage phase and recombination fraction.
}
\keyword{internal}
