% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{is.mappoly2.chitest.seq}
\alias{is.mappoly2.chitest.seq}
\title{Check if an object is of class mappoly2.chitest.seq}
\usage{
is.mappoly2.chitest.seq(x)
}
\description{
Check if an object is of class mappoly2.chitest.seq
}
