% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{is.mappoly2.geno.ord}
\alias{is.mappoly2.geno.ord}
\title{Check if an object is of class mappoly2.geno.ord}
\usage{
is.mappoly2.geno.ord(x)
}
\description{
Check if an object is of class mappoly2.geno.ord
}
