% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{is.mappoly2.sequence}
\alias{is.mappoly2.sequence}
\title{Check if an object is of class mappoly2.sequence}
\usage{
is.mappoly2.sequence(x)
}
\description{
Check if an object is of class mappoly2.sequence
}
