% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{make_sequence}
\alias{make_sequence}
\title{Create a Genetic Sequence from a Screened or Grouped Object}
\usage{
make_sequence(x, lg = NULL, ch = NULL, mrk.id.list = NULL)
}
\arguments{
\item{x}{An object of class 'screened' or 'mappoly2.group'.}

\item{lg}{A list specifying the linkage groups, if applicable.}

\item{ch}{A list specifying the chromosomes, if applicable.}

\item{mrk.id.list}{A list or vector of marker IDs to be used in the sequence.}
}
\value{
Returns a sequence object based on the provided parameters.
The sequence is constructed using internal function \code{.map_skeleton}.
}
\description{
This function generates a genetic sequence based on the provided screened or
grouped object. It allows for the creation of sequences using specific markers,
linkage groups, and chromosome information.
}
\details{
The function first checks if \code{mrk.id.list} is provided and is not null,
and uses the marker information to initiate the sequence. If \code{lg} and \code{ch}
are provided, they are used to determine the sequence structure.
The function performs various checks and validations on the input parameters
and the structure of the \code{x} object.

\if{html}{\out{<div class="sourceCode">}}\preformatted{     If `x` is of class 'mappoly2.group', the function uses data within the
     group object for sequence creation. If `lg` and `ch` are null, the function
     uses only the results of the UPGMA (Unweighted Pair Group Method with Arithmetic Mean).
}\if{html}{\out{</div>}}
}
