% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{mapping_per_group}
\alias{mapping_per_group}
\title{Multipoint analysis using Hidden Markov Models}
\usage{
mapping_per_group(
  x,
  gr,
  type = c("mds", "genome"),
  phase.conf = "all",
  rf = NULL,
  error = 0,
  verbose = FALSE,
  tol = 0.001,
  ret_H0 = FALSE
)
}
\arguments{
\item{void}{internal function}
}
\description{
Multipoint analysis using Hidden Markov Models
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
\keyword{internal}
