% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappoly.R
\docType{package}
\name{mappoly2}
\alias{mappoly2}
\title{mappoly2: A package for constructing genetic maps in autopolyploids}
\description{
mappoly2 provides ...
}
\section{mappoly main functions}{

The mappoly2 functions ...
}

