% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mappoly_to_csv_mappoly2}
\alias{mappoly_to_csv_mappoly2}
\title{Convert MAPpoly Data to CSV}
\usage{
mappoly_to_csv_mappoly2(x, path = NULL, parent.names = NULL)
}
\arguments{
\item{x}{A list of MAPpoly data objects or a single MAPpoly data object. The function
checks if \code{x} is a list, and if not, it converts it into a list.}

\item{path}{Optional; a string specifying the directory where the CSV files will be saved.
If not provided, files are saved in the current working directory.}

\item{parent.names}{Optional; a matrix of names for the parent genotypes. If not provided,
default names are generated in the format P1, P2, etc. The matrix should have two columns
and a number of rows equal to the length of \code{x}.}
}
\value{
Invisible NULL. The function is used for its side effect of writing files.
}
\description{
This function takes MAPpoly data objects and writes them to CSV files. Each object in the list
is processed individually, and a CSV file is generated for each. The user can specify the path
for saving the CSV files, and custom names for the parents in the dataset.
}
\details{
This function processes each MAPpoly data object, extracting relevant data for
CSV output. It handles missing reference and alternate sequences, and ensures that the
provided parent names match the structure of the data. The function creates one CSV file
per MAPpoly data object.
}
