% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{mds}
\alias{mds}
\title{Estimates Loci Position Using Multidimensional Scaling}
\usage{
mds(
  x,
  mrk.id,
  p = NULL,
  n = NULL,
  ndim = 2,
  weight.exponent = 2,
  verbose = TRUE
)
}
\arguments{
\item{x}{An object of class \code{mappoly2.sequence}.}

\item{mrk.id}{A vector of marker IDs to be used in the analysis.}

\item{p}{Integer; the smoothing parameter for the principal curve.
If \code{NULL} (default), this will be determined using leave-one-out cross-validation.}

\item{n}{Vector of integers or strings containing loci to be omitted from the analysis.}

\item{ndim}{Integer; the number of dimensions to be considered in the MDS procedure (default = 2).}

\item{weight.exponent}{Integer; the exponent used in the LOD score values to weight the MDS procedure (default = 2).}

\item{verbose}{Boolean; if \code{TRUE} (default), displays information about the analysis.}
}
\value{
A list containing various components related to MDS results, including the input
distance map, unconstrained MDS results, principal curve results, matrix of pairwise
distances, data frame of loci positions, total length of the segment, vector of
removed loci, scaling factor, and data frames for interpreting MDS plots.
}
\description{
This function estimates loci positions using the Multidimensional Scaling (MDS) method.
The method is adapted from the package \code{MDSmap} and is based on the approach
proposed by Preedy and Hackett (2016).
}
\details{
The function employs MDS to estimate the positions of loci on a genetic map.
It includes options for excluding certain loci, adjusting the number of dimensions,
and setting weighting factors for LOD scores. The results include detailed information
about the estimated positions and relationships between loci.
}
\references{
Preedy, K. F., & Hackett, C. A. (2016). A rapid marker ordering approach for
high-density genetic linkage maps in experimental autotetraploid populations
using multidimensional scaling. \emph{Theoretical and Applied Genetics}, 129(11),
2117-2132. \doi{10.1007/s00122-016-2761-8}
}
\author{
Marcelo Mollinari, adapted from MDSmap codes by Katharine F. Preedy
}
