% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{msg}
\alias{msg}
\title{Creates a formatted message with a horizontal rule
and specified text color.}
\usage{
msg(text, line = 1, col = "-")
}
\description{
Creates a formatted message with a horizontal rule
and specified text color.
}
\keyword{internal}
