% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{order_sequence}
\alias{order_sequence}
\title{Order Genetic Sequence in a Mapping Data Object}
\usage{
order_sequence(
  x,
  lg = NULL,
  type = c("mds", "genome"),
  p = NULL,
  n = NULL,
  ndim = 2,
  weight.exponent = 2,
  verbose = TRUE
)
}
\arguments{
\item{x}{A mapping data object that contains genetic mapping information.}

\item{lg}{An optional vector specifying the linkage groups to be ordered.
If NULL, all linkage groups in the data object are considered.}

\item{type}{The type of genetic data to be ordered, either 'mds' for
Multi-Dimensional Scaling or 'genome' for genomic data.}

\item{p}{An optional parameter for the ordering function, specific to the
method being used (e.g., MDS).}

\item{n}{An optional parameter for the ordering function, specific to the
method being used.}

\item{ndim}{The number of dimensions to be used in the ordering process.
This parameter is primarily relevant for MDS.}

\item{weight.exponent}{The exponent for weighting in the ordering process.}

\item{verbose}{A logical value indicating whether to print detailed output
during the execution of the function.}
}
\value{
The function returns the modified mapping data object with updated
order of the genetic sequence.
}
\description{
This function orders the genetic sequence within a mapping data object. It can
operate on different types of genetic data (e.g., MDS or genomic) and allows
for customization of the ordering process through various parameters.
}
\details{
The function iterates over the specified linkage groups (or all groups
if none are specified) and applies either MDS or genomic ordering
based on the 'type' parameter. Additional parameters like 'p', 'n',
and 'ndim' are used to fine-tune the ordering process.
}
