% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasing.R
\name{pairwise_phasing_per_group}
\alias{pairwise_phasing_per_group}
\title{Phasing based on pairwise recombination fraction estimation}
\usage{
pairwise_phasing_per_group(
  x,
  gr,
  type = c("mds", "genome"),
  thresh.LOD.ph = 3,
  thresh.LOD.rf = 3,
  thresh.rf = 0.5,
  max.conf.btnk.p1 = 1,
  max.conf.btnk.p2 = max.conf.btnk.p1,
  verbose = TRUE
)
}
\arguments{
\item{void}{internal function}
}
\description{
Phasing based on pairwise recombination fraction estimation
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
