% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_rf.R
\name{pairwise_rf}
\alias{pairwise_rf}
\title{Pairwise Two-Point Analysis}
\usage{
pairwise_rf(
  x,
  mrk.scope = c("all", "per.chrom", "chrom"),
  chrom = NULL,
  ncpus = 1L,
  verbose = TRUE,
  tol = .Machine$double.eps^0.25
)
}
\arguments{
\item{x}{An object of class \code{mappoly2.data}}

\item{mrk.scope}{Specifies the range of markers for which the pairwise recombination
fractions will be calculated. Acceptable values are "all", "per.chrom", and "chrom".
See details for more information.}

\item{chrom}{Specifies the particular chromosome for which the
pairwise recombination fractions will be calculated. This argument is required when
the \code{mrk.scope} argument is set to \code{"chrom"}.}

\item{ncpus}{Number of parallel processes (cores) to use (default = 1)}

\item{verbose}{Logical; if TRUE, progress messages will be printed.}

\item{tol}{Desired accuracy. See \code{optimize()} for details}
}
\value{
An updated object of class \code{mappoly2.data} with a slot called 'pairwise.rf' containing:
\itemize{
\item \code{"rec.mat"} the recombination fraction matrix
\item \code{"lod.mat"} the LOD Scrore associated to the recombination fraction matrix
\item \code{"lod.ph.mat"} the LOD Scrore associated to the second most likely linkage phase configuration
\item \code{"Sh.p1" and "Sh.p2"} the number of homologs that share the alternate alleles for the estimated linkage phase configurationns for parents 1 and 2, respectively.
}
}
\description{
Performs a pairwise two-point analysis for selected markers in a dataset.
The function estimates the recombination fraction for all possible linkage phase
configurations and their respective LOD Scores, returning the most likely one.
}
\details{
The \code{mrk.scope} argument allows for the customization of the analysis scope. Options are:
\itemize{
\item \code{"all"}: Analyze all markers in the dataset.
\item \code{"per.chrom"}: Analyze markers within each chromosome.
\item \code{"chrom"}: Analyze markers in a specific chromosome. Requires specifying the chromosome.
}

Additional arguments (\code{chrom} or \code{sequence}) must be specified when using
\code{"chrom"} or \code{"seq"} options, respectively.
}
\examples{

  dat <- subset(B2721, perc = .1)
  dat <- filter_data(dat, mrk.thresh = .08, ind.thresh = 0.06)
  dat <- pairwise_rf(dat, mrk.scope = "chrom", chrom = "ch10")

}
