% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_lg_and_type}
\alias{parse_lg_and_type}
\title{Parse Linkage Group and Type}
\usage{
parse_lg_and_type(x, lg = NULL, type = c("mds", "genome", "custom"))
}
\arguments{
\item{x}{An object of class mappoly2.sequence. This object should contain SNP data
along with associated maps.}

\item{lg}{A character or numeric vector specifying the linkage groups. If a character
vector, it should correspond to the names of the maps in the mappoly2.sequence object.
If numeric, it should be the indices of the maps.}

\item{type}{A character vector specifying the type of SNP order.
Possible values are "mds", "genome", or "custom". Default is c("mds", "genome", "custom").
The function will match the provided type argument with these options.}
}
\value{
A list containing two elements: 'lg' and 'type'. The 'lg' element contains the
linkage group information, either as a character or numeric vector, as provided by the user.
The 'type' element contains the matched SNP order type.
}
\description{
This function parses the linkage group and type from a given mappoly2.sequence object.
It validates the input and extracts the specified linkage group and type.
}
\keyword{internal}
