% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasing.R
\name{phase_remaining}
\alias{phase_remaining}
\title{Phasing remaining markers based on pairwise recombination fraction estimation and multilocus estimation}
\usage{
phase_remaining(
  input.seq,
  input.twopt,
  thresh.LOD.ph = 3,
  thresh.LOD.rf = 3,
  thresh.rf = 0.5,
  max.phases = 3,
  tol = 0.001,
  verbose = TRUE
)
}
\arguments{
\item{void}{internal function}
}
\description{
Phasing remaining markers based on pairwise recombination fraction estimation and multilocus estimation
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
