% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.mappoly2.consensus.map}
\alias{plot.mappoly2.consensus.map}
\title{Plot Consensus Map}
\usage{
\method{plot}{mappoly2.consensus.map}(x, only.consensus = FALSE, col = "lightgray", ...)
}
\arguments{
\item{x}{A list containing elements of 'mappoly2.prepared.integrated.data' class,
which includes both individual maps and consensus map data.}

\item{only.consensus}{Logical, if TRUE, only the consensus map is plotted;
if FALSE, individual population maps are included (default is FALSE).}

\item{col}{The color used for plotting the consensus map markers when
'only.consensus' is TRUE (default is "lightgray").}

\item{...}{Additional arguments, not used in this method}
}
\value{
A ggplot object representing the plotted genetic map(s).
}
\description{
This function plots consensus genetic maps along with individual population maps.
It allows the option to plot only the consensus map or include individual population maps.
}
