% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.mappoly2.data}
\alias{plot.mappoly2.data}
\title{Plot Data from mappoly2.data Object}
\usage{
\method{plot}{mappoly2.data}(x, type = c("rf", "screened", "density", "raw"), chrom = NULL, ...)
}
\arguments{
\item{x}{A \code{mappoly2.data} object containing genetic mapping data. The object can also be of class "screened" and "pairwise.rf".}

\item{type}{A character string specifying the type of plot to generate. It can be one of "rf" (recombination frequency), "screened", "density", or "raw". Default is "rf".}

\item{chrom}{Optional; a vector of chromosome numbers or names to subset the data before plotting. If \code{NULL}, all chromosomes are considered.}

\item{...}{Additional arguments, not used in this method}
}
\value{
The function does not return a value but generates a plot.
}
\description{
Visualizes data from a \code{mappoly2.data} object, which may also have "screened" and "pairwise.rf" classes. This function provides various plotting options including recombination frequency matrices, screened data, marker density, and raw data, depending on the type of data available in the object and the selected options.
}
\details{
The function can visualize data in four different ways based on the \code{type} parameter:
\itemize{
\item "rf": Plots a recombination frequency matrix, available for objects with the "pairwise.rf" class.
\item "screened": Shows screened data for objects with the "screened" class.
\item "density": Generates a density plot of markers, requiring genome position information.
\item "raw": Displays raw data from the \code{mappoly2.data} object.
The function handles data selection based on the provided chromosome information (\code{chrom}).
}
}
\examples{
plot(B2721)

}
