% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_consensus_haplo}
\alias{plot_consensus_haplo}
\title{Plot Consensus Map with Homolog Probabilities}
\usage{
plot_consensus_haplo(x, lg = 1, ind = 1, ...)
}
\arguments{
\item{x}{A list containing consensus map data, specifically \code{consensus.map}, \code{ph}, \code{pedigree}, and \code{haploprob} structures.}

\item{lg}{An integer specifying the linkage group to plot. Defaults to 1.}

\item{ind}{An identifier for the individual to plot. This can be either a numeric index or a character string corresponding to the individual's name. If it is a character string, the function matches it with the rownames of the pedigree. Defaults to 1.}

\item{...}{Additional arguments to pass on to the plotting function.}
}
\value{
A ggplot object representing the density plot of homolog probabilities for the specified individual across map positions.
}
\description{
This function generates a density plot for homolog probabilities across map positions
for a specified individual within a given linkage group. It uses haplotype probability data
from a consensus map to create the plot, with different colors representing different parents.
}
