% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_genome_vs_map}
\alias{plot_genome_vs_map}
\title{Physical versus genetic distance}
\usage{
plot_genome_vs_map(
  x,
  type = c("mds", "genome"),
  same.ch.lg = FALSE,
  alpha = 1/5,
  size = 3
)
}
\arguments{
\item{same.ch.lg}{Logical. If \code{TRUE} displays only the scatterplots between the
chromosomes and linkage groups with the same number. Default is \code{FALSE}.}

\item{alpha}{transparency factor for SNPs points}

\item{size}{size of the SNP points}

\item{w}{A list or a single object of class \code{mappoly.map}}

\item{phase.config}{A vector containing which phase configuration should be
plotted. If \code{'best'} (default), plots the configuration
with the highest likelihood for all elements in \code{'w'}}
}
\description{
This function plots scatterplot(s) of physical distance (in Mbp) versus the genetic
distance (in cM). Map(s) should be passed as a single object or a list of objects
of class \code{mappoly.map}.
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
