% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_haplotypes}
\alias{plot_haplotypes}
\title{Plot Haplotype Probabilities for Genetic Maps}
\usage{
plot_haplotypes(
  x,
  lg = NULL,
  ind = 1,
  type = c("mds", "genome"),
  parent = c("p1p2", "p1", "p2")
)
}
\arguments{
\item{x}{An object representing genetic mapping data, typically of a class storing genetic information.}

\item{lg}{Optional vector specifying the linkage group indices to plot.
If NULL, the first linkage group is processed.}

\item{ind}{Index or name of the individual for which haplotypes are to be plotted.
Should be a single numeric index or a character name.}

\item{type}{Character vector indicating the type of map to process, either "mds" or "genome".}

\item{parent}{Character vector specifying the parent or parents to be considered
in the haplotype probability visualization. Options are "p1p2", "p1", and "p2".}
}
\value{
A ggplot object representing the plotted haplotype probabilities.
}
\description{
This function plots haplotype probabilities for specified linkage groups in a genetic mapping dataset.
It supports visualization of haplotype probabilities for individual or multiple linkage groups.
}
