% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_map_list}
\alias{plot_map_list}
\title{Plot a List of Genetic Maps}
\usage{
plot_map_list(
  x,
  horiz = TRUE,
  type = c("mds", "genome"),
  parent = c("p1p2", "p1", "p2"),
  col = mp2_pal(length(x$maps))
)
}
\arguments{
\item{x}{A mappoly2 sequence object containing genetic map data.}

\item{horiz}{Logical; if TRUE, the maps are plotted horizontally, otherwise vertically.}

\item{type}{The type of genetic maps to be plotted, either "mds" or "genome".}

\item{parent}{Specifies which parent's data to use in the visualization.
Options are "p1p2" (both parents), "p1" (first parent), or "p2" (second parent).}

\item{col}{The color used for plotting the maps, default is "lightgray".
Can be a vector of colors to apply different colors to each map.}
}
\value{
The function does not return a value but generates a plot or a series of plots
representing the genetic maps. It invisibly returns a data frame containing
marker positions and linkage group information.
}
\description{
This function visualizes a list of genetic maps from a mappoly2 sequence object. It supports both horizontal and vertical orientations and allows for customization of plot colors.
}
\details{
The function iterates over the linkage groups in the provided mappoly2 sequence
object and creates a plot (or a series of plots) showing the genetic map(s)
for the specified linkage groups. The plot orientation can be set to either
horizontal or vertical, and the color of the plots can be customized.
}
