% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{plot_mds}
\alias{plot_mds}
\title{Plot Multi-Dimensional Scaling (MDS) Maps}
\usage{
plot_mds(
  x,
  lg = 1,
  D1lim = NULL,
  D2lim = NULL,
  D3lim = NULL,
  displaytext = FALSE
)
}
\arguments{
\item{x}{A genetic mapping data object containing MDS information.}

\item{lg}{A numeric value specifying the linkage group to be visualized.
Defaults to the first linkage group.}

\item{D1lim}{Optional range for the first dimension of the MDS plot.}

\item{D2lim}{Optional range for the second dimension of the MDS plot.}

\item{D3lim}{Optional range for the third dimension of the MDS plot, applicable
only for three-dimensional plots.}

\item{displaytext}{A logical value indicating whether to display text labels
on the MDS plot. Defaults to \code{FALSE}.}
}
\value{
The function does not return a value but generates a plot of the MDS
analysis results for the specified linkage group.
}
\description{
This function visualizes the results of Multi-Dimensional Scaling (MDS) analysis
on genetic mapping data. It provides options to display either two-dimensional
or three-dimensional MDS plots based on the number of dimensions in the MDS object.
}
\details{
Depending on the dimensionality of the MDS object (\code{ndim}), this
function calls either \code{plot_pcmap} for two-dimensional data or
\code{plot_pcmap3d} for three-dimensional data. It visualizes the spatial
arrangement of markers in the specified dimensions, providing insights
into the genetic structure.
}
