% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_mds_vs_genome}
\alias{plot_mds_vs_genome}
\title{Plot MDS vs. Genome}
\usage{
plot_mds_vs_genome(x, alpha = 1/2, size = 2)
}
\arguments{
\item{x}{A \code{mappoly2.sequence} object containing marker and map data. This object typically results from mapping or marker analysis processes in the \code{mappoly2} package.}

\item{alpha}{The transparency level of the points in the plot, with 1 being fully opaque and 0 being fully transparent. Defaults to 1/2.}

\item{size}{The size of the points in the plot. Defaults to 2.}
}
\value{
A ggplot object representing the MDS versus genome position plot. Each linkage group is displayed in a separate panel.
}
\description{
This function generates a plot comparing MDS (Multi-Dimensional Scaling) positions to genome positions for genetic markers in a \code{mappoly2.sequence} object. It uses functions from the \code{mappoly2} package to extract marker data and \code{ggplot2} for visualization.
}
