% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_rf_matrix}
\alias{plot_rf_matrix}
\title{Plot Recombination Fraction Matrices for Genetic Maps}
\usage{
plot_rf_matrix(x, lg = NULL, type = c("mds", "genome", "custom"), fact = 1)
}
\arguments{
\item{x}{A \code{mappoly2.data} object that contains recombination fraction information.}

\item{lg}{Optional vector specifying the linkage groups for which the recombination
fraction matrices should be plotted. If NULL, matrices for all linkage
groups are plotted.}

\item{type}{The type of genetic data to be visualized. Can be 'mds', 'genome',
or 'custom'. This parameter determines how the matrices are processed
and displayed.}

\item{fact}{A numeric factor used for scaling or aggregating the matrix data.
Defaults to 1 (no scaling).}
}
\value{
The function does not return a value but generates a series of plots,
each representing the recombination fraction matrix for a specified linkage
group.
}
\description{
This function visualizes the recombination fraction matrices for genetic maps.
It allows the user to plot these matrices for different types of genetic data
(e.g., MDS, genome, custom) and for specified linkage groups.
}
\details{
The function processes the genetic mapping data based on the specified
'type' and 'lg' parameters. It then uses \code{plot_rf_matrix_one} to plot
individual matrices for each linkage group. The visualization helps in
understanding the recombination patterns and genetic distances between
markers.
}
