% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_shared_markers}
\alias{plot_shared_markers}
\title{Plot Shared Markers in Genetic Maps}
\usage{
plot_shared_markers(x)
}
\arguments{
\item{x}{An object of class 'mappoly2.prepared.integrated.data'.
It should contain individual maps from which shared markers are to be identified.}
}
\value{
An Euler diagram plot showing shared markers across the individual maps in \code{x}.
}
\description{
This function visualizes shared markers across multiple genetic maps
in a 'mappoly2.prepared.integrated.data' object. It uses Euler diagrams to represent the intersection of markers.
}
\details{
The function extracts markers from each map within the provided 'mappoly2.prepared.integrated.data'
object and then plots an Euler diagram to show the intersections (shared markers) among these maps.
This visualization helps in understanding the overlap of genetic information across different maps.
}
