% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{potato}
\alias{potato}
\title{Autotetraploid Potato Dataset}
\format{
A list of class \code{mappoly2.data} containing the following components:
\describe{
\item{ploidy.p1}{The ploidy level of parent P1 (Atlantic): 4.}
\item{ploidy.p2}{The ploidy level of parent P2 (B1829-5): 4.}
\item{n.ind}{The total number of individuals: 156.}
\item{n.mrk}{The number of unique markers after filtering for redundancy: 6511.}
\item{ind.names}{Character vector containing the names of the individuals.}
\item{mrk.names}{Character vector containing the names of the markers.}
\item{name.p1}{The name of parent P1: Atlantic.}
\item{name.p2}{The name of parent P2: B1829-5.}
\item{dosage.p1}{Named integer vector with the dosage for parent P1 across \code{n.mrk} markers.}
\item{dosage.p2}{Named integer vector with the dosage for parent P2 across \code{n.mrk} markers.}
\item{chrom}{Named character vector indicating the chromosome each marker belongs to.}
\item{genome.pos}{Named numeric vector with the physical positions of the markers.}
\item{ref}{Character vector of reference alleles.}
\item{alt}{Character vector of alternate alleles.}
\item{geno.dose}{Numeric matrix containing dosage information for each marker (rows) and individual (columns).}
\item{redundant}{Data frame listing \code{kept} and \code{removed} markers filtered for redundancy.}
\item{QAQC.values}{A list containing quality assurance and quality control statistics:
\describe{
\item{Marker Statistics}{Data frame with statistics such as:
\code{miss} (missing data rate),
\code{chisq.pval} (chi-squared test p-value),
and \code{read.depth} (read depth).}
\item{Individual Statistics}{Data frame with statistics for each individual:
\code{miss} (missing data rate),
and \code{full.sib} (indicator of misclassified individuals not belonging
to the cross, generated using \code{\link[mappoly]{filter_individuals}}).}
}
}
}
}
\usage{
potato
}
\description{
This dataset contains 156 offspring derived from a cross between two tetraploid
potato varieties, Atlantic and B1829-5. It was genotyped using the SolCAP Infinium
8303 potato array, and the genomic order of SNPs follows the Solanum tuberosum
genome version 4.03. Genotype calling was performed with the ClusterCall R package.
The original dataset is available in \href{https://doi.org/10.1038/s41437-021-00416-x}{Pereira et al. (2021)}.
}
\keyword{datasets}
