% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{prepare_map}
\alias{prepare_map}
\title{prepare maps for plot}
\usage{
prepare_map(
  x,
  ploidy.p1,
  ploidy.p2,
  name.p1,
  name.p2,
  dosage.p1,
  dosage.p2,
  alt = NULL,
  ref = NULL
)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
prepare maps for plot
}
\keyword{internal}
