% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{prepare_to_integrate}
\alias{prepare_to_integrate}
\title{Prepare Genetic Map Data for Integration}
\usage{
prepare_to_integrate(x, lg = NULL, type = c("genome", "mds"), verbose = TRUE)
}
\arguments{
\item{x}{A list of genetic map objects.}

\item{lg}{Linkage group to be analyzed, default is NULL.}

\item{type}{Type of the map, either "genome" or "mds" (default is "genome").}

\item{verbose}{Logical, indicating whether to show detailed messages (default is TRUE).}
}
\value{
A list of prepared data for building an integrated genetic map.
}
\description{
This function prepares multiple biparental genetic maps for integration into
a unified multi-population genetic map. It primarily utilizes an internal
function \code{match_homologs} to match and rearrange homologs across shared
parents in different populations, ensuring consistency in the integrated
map estimation. Markers with inconsistencies, such as varying dosages or phasing,
are removed. The function currently supports only genome-ordered maps
and operates on a single chromosome.
}
