% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_biparental_geno_csv}
\alias{read_biparental_geno_csv}
\title{Read Genetic Marker Data from a CSV File}
\usage{
read_biparental_geno_csv(
  file.in,
  ploidy.p1,
  ploidy.p2 = ploidy.p1,
  name.p1 = NULL,
  name.p2 = NULL,
  filter.non.conforming = TRUE,
  filter.redundant = TRUE,
  verbose = TRUE,
  reorder = TRUE
)
}
\arguments{
\item{file.in}{A character string specifying the name or full path to the input CSV file.}

\item{ploidy.p1}{The ploidy level of parent 1.}

\item{ploidy.p2}{The ploidy level of parent 2 (defaults to \code{ploidy.p1}).}

\item{name.p1}{The name of parent 1.}

\item{name.p2}{The name of parent 2.}

\item{filter.non.conforming}{Logical value indicating whether to convert non-conforming data points
(e.g., double reduction) to \code{NA} (default is \code{TRUE}).}

\item{filter.redundant}{Logical value indicating whether to remove redundant markers during map construction,
retaining annotations for export in the final map (default is \code{TRUE}).}

\item{verbose}{Logical value indicating whether to display progress updates (default is \code{TRUE}).}
}
\value{
An object of class \code{mappoly2.data}, which is a list containing the following components:
\describe{
\item{ploidy.p1}{Ploidy level of parent 1.}
\item{ploidy.p2}{Ploidy level of parent 2.}
\item{n.ind}{Number of individuals in the population.}
\item{n.mrk}{Total number of markers.}
\item{ind.names}{Names or identifiers of the individuals.}
\item{mrk.names}{Names or identifiers of the genetic markers.}
\item{name.p1}{Name or identifier of parent 1.}
\item{name.p2}{Name or identifier of parent 2.}
\item{dosage.p1}{Dosage for parent 1.}
\item{dosage.p2}{Dosage for parent 2.}
\item{chrom}{Chromosome numbers for all markers.}
\item{genome.pos}{Physical positions of the markers within the genome.}
\item{ref}{Reference alleles for the markers.}
\item{alt}{Alternate alleles for the markers.}
\item{all.mrk.depth}{Depth of coverage for all markers (NULL when using CSV input files).}
\item{geno.dose}{Matrix of dosages for each marker (rows) and each individual (columns).}
\item{redundant}{List of non-redundant markers and their equivalent redundant markers if
\code{filter.redundant} is \code{TRUE}.}
\item{QAQC.values}{List containing quality assurance and quality control values:
\describe{
\item{markers}{Data frame with statistics for each marker, including \code{miss}
(missing data rate) and \code{chisq.pval} (chi-squared test p-value).}
\item{individuals}{Data frame with statistics for each individual, including \code{miss}
(missing data rate) and \code{full.sib} (indicator of non-belonging to the analyzed
bi-parental cross).}
}
}
}
}
\description{
Reads genetic marker data from a CSV file and returns an object of class \code{mappoly2.data}.
}
\details{
The CSV file should have markers in rows, with the first row serving as the header.
The first seven columns must contain the following information, in order:
\enumerate{
\item Marker names
\item Dosage in parent 1
\item Dosage in parent 2
\item Chromosome information (e.g., chromosome, scaffold, contig)
\item Position of the marker within the sequence
\item Alternate allele (if available)
\item Reference allele (if available)
}
If allele information is not available, the values should be set to \code{NA}.
The remaining columns should contain the dosage for each individual in the full-sib population.
Refer to the \code{Examples} section for a tetraploid example.
}
\examples{
\donttest{
# Read a tetraploid dataset from a CSV file
tempfl <- list.files(system.file('extdata', package = 'mappoly2'), full.names = TRUE)
alfalfa.bc <- read_biparental_geno_csv(
  file.in = tempfl,
  ploidy.p1 = 4,
  name.p1 = "I195",
  name.p2 = "F1.85.209"
)
print(alfalfa.bc, detailed = TRUE)
plot(alfalfa.bc)
}

}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
