% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_biparental_geno_vcf}
\alias{read_biparental_geno_vcf}
\title{Read Genetic Marker Data from a VCF File}
\usage{
read_biparental_geno_vcf(
  file.in,
  ploidy.p1,
  ploidy.p2 = ploidy.p1,
  name.p1,
  name.p2,
  name.offspring = NULL,
  filter.non.conforming = TRUE,
  filter.redundant = TRUE,
  verbose = TRUE,
  min.gt.depth = 0,
  min.av.depth = 0,
  max.missing = 1
)
}
\arguments{
\item{file.in}{A character string specifying the name or full path of the input VCF file.}

\item{ploidy.p1}{The ploidy level of parent 1.}

\item{ploidy.p2}{The ploidy level of parent 2 (defaults to \code{ploidy.p1}).}

\item{name.p1}{The name of parent 1.}

\item{name.p2}{The name of parent 2.}

\item{name.offspring}{A character vector containing the names of the offspring (defaults to all
individuals except the parents).}

\item{filter.non.conforming}{Logical value indicating whether to convert non-conforming data points
to \code{NA} (default is \code{TRUE}).}

\item{filter.redundant}{Logical value indicating whether to remove redundant markers during map
construction, retaining annotations for export in the final map (default is \code{TRUE}).}

\item{verbose}{Logical value indicating whether to display progress information (default is \code{TRUE}).}

\item{min.gt.depth}{Minimum genotype depth to retain information; values below this threshold are
replaced with \code{NA} (default is 0).}

\item{min.av.depth}{Minimum average depth to retain markers (default is 0).}

\item{max.missing}{Maximum proportion of missing data allowed to retain markers (range from 0 to 1;
default is 1).}
}
\value{
An object of class \code{mappoly2.data}, which is a list containing the following components:
\describe{
\item{ploidy.p1}{Ploidy level of parent 1.}
\item{ploidy.p2}{Ploidy level of parent 2.}
\item{n.ind}{Number of individuals.}
\item{n.mrk}{Total number of markers.}
\item{ind.names}{Names of the individuals.}
\item{mrk.names}{Names of the genetic markers.}
\item{name.p1}{Name of parent 1.}
\item{name.p2}{Name of parent 2.}
\item{dosage.p1}{Dosage information for parent 1.}
\item{dosage.p2}{Dosage information for parent 2.}
\item{chrom}{Chromosome numbers for all markers.}
\item{genome.pos}{Physical positions of the markers within the genome.}
\item{ref}{Reference alleles for the markers.}
\item{alt}{Alternate alleles for the markers.}
\item{all.mrk.depth}{Depth of coverage for all markers.}
\item{geno.dose}{Matrix of dosages for each marker (rows) and each individual (columns).}
\item{redundant}{List of non-redundant markers and their equivalent redundant markers if
\code{filter.redundant} is \code{TRUE}.}
\item{QAQC.values}{Quality assurance and quality control values:
\describe{
\item{markers}{Data frame with statistics for each marker, including \code{miss}
(missing data rate) and \code{chisq.pval} (chi-squared test p-value).}
\item{individuals}{Data frame with statistics for each individual, including \code{miss}
(missing data rate) and \code{full.sib} (indicator of non-belonging to the analyzed
bi-parental cross).}
}
}
}
}
\description{
Reads genetic marker data from a Variant Call Format (VCF) file and returns an object of class
\code{mappoly2.data}.
}
\details{
Supports VCF files of version 4.0 or higher.
}
\examples{
\donttest{
# Hexaploid sweetpotato: Subset of chromosome 3
fl <- "https://github.com/mmollina/MAPpoly_vignettes/raw/master/data/sweet_sample_ch3.vcf.gz"
tempfl <- tempfile(pattern = 'chr3_', fileext = '.vcf.gz')
download.file(fl, destfile = tempfl)
dat.dose.vcf <- read_biparental_geno_vcf(
  file.in = tempfl,
  ploidy.p1 = 6,
  name.p1 = "PARENT1",
  name.p2 = "PARENT2",
  min.av.depth = 20
)
print(dat.dose.vcf)
plot(dat.dose.vcf)
}

}
\references{
Gabriel Gesteira, \email{gdesiqu@ncsu.edu}; Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
