% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_geno_csv}
\alias{read_geno_csv}
\title{Data Input in CSV Format}
\usage{
read_geno_csv(
  file.in,
  ploidy.p1,
  ploidy.p2 = ploidy.p1,
  name.p1 = NULL,
  name.p2 = NULL,
  filter.non.conforming = TRUE,
  filter.redundant = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{file.in}{A character string with the name of (or full path to)
the input file containing the data to be read.}

\item{ploidy.p1}{Ploidy level of parent 1.}

\item{ploidy.p2}{Ploidy level of parent 2.}

\item{name.p1}{Name of parent 1.}

\item{name.p2}{Name of parent 2.}

\item{filter.non.conforming}{If \code{TRUE} (default), data points with
unexpected genotypes (e.g., double reduction) are converted to 'NA'.
See the \code{\link[mappoly]{segreg_poly}} function for information on
expected classes and their respective frequencies.}

\item{filter.redundant}{Logical. If \code{TRUE} (default), removes redundant
markers during map construction, keeping them annotated to export to the
final map.}

\item{verbose}{If \code{TRUE} (default), shows the current progress; if
\code{FALSE}, no output is produced.}

\item{x}{An object of the class \code{mappoly.sequence}.}

\item{thresh.line}{Position of a threshold line for p-values of the segregation test (default = \code{0.05/n.mrk}).}

\item{...}{Currently ignored.}
}
\value{
An object of the class \code{mappoly2} which contains a list with
the following components:

\item{ploidy.p1}{Ploidy level of the first parent.}
\item{ploidy.p2}{Ploidy level of the second parent.}
\item{n.ind}{Number of individuals.}
\item{n.mrk}{Total number of markers.}
\item{ind.names}{Names or identifiers of the individuals.}
\item{mrk.names}{Names or identifiers of the genetic markers.}
\item{name.p1}{Name or identifier of the first parent.}
\item{name.p2}{Name or identifier of the second parent.}
\item{dosage.p1}{The dosage for the first parent.}
\item{dosage.p2}{The dosage for the second parent.}
\item{chrom}{Chromosome numbers for all markers.}
\item{genome.pos}{Physical positions on the genome for the genetic markers.}
\item{seq.ref}{The reference DNA sequence data for the genetic markers.}
\item{seq.alt}{The alternate DNA sequence data for the genetic markers.}
\item{all.mrk.depth}{Represents the depth of coverage for all genetic markers.
NULL when using CSV input files.}
\item{geno.dose}{A matrix containing the dosage for each marker (rows) for
each individual (columns).}
\item{kept}{A list of all non-redundant markers if \code{filter.redundant = TRUE}.}
\item{filter.correspondence}{A list of all non-redundant markers and their equivalence
to the redundant ones if \code{filter.redundant = TRUE}.}
}
\description{
Reads a comma-separated values (CSV) data file containing genetic marker data.
This function returns an object of the class \code{mappoly2}.
}
\details{
The CSV file should have rows representing markers, with the first row being
used as a header. The first seven columns should contain the marker names, the
dosages in parents 1 and 2, chromosome information (i.e., chromosome,
scaffold, contig, etc.), the position of the marker within the sequence, and
the alternate and reference alleles, if available. If not available, the values
should be filled with NA.
The remaining columns should contain the dosage of the full-sib population.
For a tetraploid example of such a file, see the \code{Examples} section.
}
\examples{
\donttest{
tempfl <- list.files(system.file('extdata', package = 'mappoly2'),
                     full.names = TRUE)
SolCAP.dose <- read_geno_csv(file.in = tempfl,
                             ploidy.p1 = 4,
                             name.p1 = "Atlantic",
                             name.p2 = "B1829-5")
print(SolCAP.dose, detailed = TRUE)
plot(SolCAP.dose)
}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
