% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_vcf}
\alias{read_vcf}
\title{Read VCF File for Genetic Marker Data}
\usage{
read_vcf(
  file.in,
  ploidy.p1,
  ploidy.p2 = ploidy.p1,
  name.p1,
  name.p2,
  name.offspring = NULL,
  filter.non.conforming = TRUE,
  filter.redundant = TRUE,
  verbose = TRUE,
  min.gt.depth = 0,
  min.av.depth = 0,
  max.missing = 1
)
}
\arguments{
\item{file.in}{A character string specifying the name or full path of the input file containing the data in VCF format.}

\item{ploidy.p1}{An integer indicating the ploidy level of parent 1.}

\item{ploidy.p2}{An integer indicating the ploidy level of parent 2. Defaults to the value of \code{ploidy.p1}.}

\item{name.p1}{A character string containing the name of parent 1.}

\item{name.p2}{A character string containing the name of parent 2.}

\item{name.offspring}{A character vector containing the names of the offspring. Defaults to \code{NULL}.}

\item{filter.non.conforming}{Logical, if \code{TRUE} (default), converts data points with unexpected genotypes to 'NA'. See \code{\link[mappoly]{segreg_poly}} for details on expected classes and frequencies.}

\item{filter.redundant}{Logical, if \code{TRUE} (default), removes redundant markers during map construction, keeping them annotated for export to the final map.}

\item{verbose}{Logical, if \code{TRUE} (default), displays progress information.}

\item{min.gt.depth}{An integer specifying the minimum genotype depth to retain information. Values below \code{min.gt.depth} are replaced with \code{NA} (default = 0).}

\item{min.av.depth}{An integer specifying the minimum average depth to retain markers (default = 0).}

\item{max.missing}{A numeric value specifying the maximum proportion of missing data allowed to retain markers (range = 0-1; default = 1).}
}
\value{
An object of class \code{mappoly2.data} containing the following components:
\describe{
\item{ploidy.p1}{Ploidy level of parent 1.}
\item{ploidy.p2}{Ploidy level of parent 2.}
\item{n.ind}{Number of individuals.}
\item{n.mrk}{Total number of markers.}
\item{ind.names}{Names of the individuals.}
\item{mrk.names}{Names of the genetic markers.}
\item{name.p1}{Name of parent 1.}
\item{name.p2}{Name of parent 2.}
\item{dosage.p1}{Dosage information for parent 1.}
\item{dosage.p2}{Dosage information for parent 2.}
\item{chrom}{Chromosome numbers for all markers.}
\item{genome.pos}{Physical positions of the genetic markers on the genome.}
\item{ref}{Reference DNA sequence data for the genetic markers.}
\item{alt}{Alternate DNA sequence data for the genetic markers.}
\item{all.mrk.depth}{Depth of coverage for all genetic markers.}
\item{geno.dose}{A matrix of dosage information for each marker (rows) and each individual (columns).}
\item{redundant}{A list of non-redundant markers and their equivalent redundant markers if \code{filter.redundant} is TRUE.}
\item{QAQC.values}{Quality assurance and quality control values:
\describe{
\item{markers}{A data frame with statistics for each marker, including \code{miss} (missing data rate), \code{chisq.pval} (chi-squared test p-value), and \code{read.depth} (read depth).}
\item{individuals}{A data frame with statistics for each individual, including \code{miss} (missing data rate) and \code{full.sib} (indicator of non-belonging to the analyzed bi-parental cross, generated by \code{\link[mappoly]{filter_individuals}}).}
}
}
}
}
\description{
This function reads genetic marker data from a Variant Call Format (VCF) file and returns an object of class \code{mappoly2.data}.
}
\details{
This function supports VCF files of version 4.0 or higher.
}
\examples{
\donttest{
## Hexaploid sweetpotato: Subset of chromosome 3
fl <- "https://github.com/mmollina/MAPpoly_vignettes/raw/master/data/sweet_sample_ch3.vcf.gz"
tempfl <- tempfile(pattern = 'chr3_', fileext = '.vcf.gz')
download.file(fl, destfile = tempfl)
dat.dose.vcf <- read_vcf(file.in = tempfl, ploidy.p1 = 6, name.p1 = "PARENT1", name.p2 = "PARENT2", min.av.depth = 20)
print(dat.dose.vcf)
plot(dat.dose.vcf)
}

}
\references{
Gabriel Gesteira, \email{gdesiqu@ncsu.edu}, Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
