% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_edits.R
\name{remove_gap_markers}
\alias{remove_gap_markers}
\title{Remove Markers Causing Gaps in Genetic Maps}
\usage{
remove_gap_markers(
  x,
  lg = NULL,
  type = c("mds", "genome"),
  parent = c("p1p2", "p1", "p2"),
  gap.threshold = 5,
  size.rem.cluster = 1,
  ncpus = 1,
  reestimate.hmm.map = TRUE,
  recompute.haplotype.prob = TRUE,
  verbose = TRUE,
  tol = 0.001,
  error = NULL
)
}
\arguments{
\item{x}{A genetic mapping object, typically of a class "mappoly2.sequence".}

\item{lg}{Optional vector specifying the linkage group indices to process.
If NULL, all linkage groups in \code{x} are processed.}

\item{type}{Character vector indicating the type of map to process, either "mds" or "genome".}

\item{parent}{Character vector specifying the parent or parents to be considered
in the marker removal process. Options are "p1p2" (both parents),
"p1" (first parent), and "p2" (second parent).}

\item{gap.threshold}{The threshold for identifying significant gaps in the map.
Markers causing gaps larger than this threshold will be considered for removal.}

\item{size.rem.cluster}{Minimum number of consecutive markers to retain.
Segments with fewer markers than this threshold will be removed.}

\item{ncpus}{The number of CPU cores to use for parallel processing. Defaults to 1.}

\item{reestimate.hmm.map}{If TRUE, re-estimates the hidden Markov model for the map after marker removal.}

\item{recompute.haplotype.prob}{If TRUE, recalculates haplotype probabilities after marker removal.}

\item{verbose}{Logical; if \code{TRUE}, function prints messages during execution (default is \code{TRUE}).}

\item{tol}{Tolerance level for the mapping algorithm.}

\item{error}{Optional; error rate to be used in the hidden Markov model estimation.
If NULL, it uses the error rate from the input object.}
}
\value{
An updated genetic mapping object with problematic markers removed and, if specified,
re-estimated HMM and recalculated haplotype probabilities.
}
\description{
This function identifies and removes markers from genetic maps that potentially cause gaps.
These gaps may be due to misplaced markers, incorrect phasing, or markers with anomalous behavior.
It aims to improve the quality and accuracy of the map by excluding such problematic markers.
}
\details{
The function analyzes the genetic map to identify markers causing gaps
larger than the specified threshold. It then removes these markers and, optionally,
re-estimates the hidden Markov model and recalculates haplotype probabilities
to refine the map.
}
