% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rev_map}
\alias{rev_map}
\title{Reverse the Order of a Genetic Map}
\usage{
rev_map(
  x,
  lg,
  type = c("mds", "genome", "custom"),
  parent = c("p1p2", "p1", "p2")
)
}
\arguments{
\item{x}{A \code{mappoly} object containing the genetic map(s) to be reversed.}

\item{lg}{An integer or vector specifying the linkage group(s) to be reversed.}

\item{type}{Character vector indicating the type of genetic map to be reversed.
Options include "mds", "genome", or "custom". Default is c("mds", "genome", "custom").}

\item{parent}{A character vector specifying the parent or parents to be considered
in the mapping process. Options are "p1p2" (both parents), "p1" (first parent),
and "p2" (second parent). Default is c("p1p2", "p1", "p2").}
}
\value{
Returns the \code{mappoly} object with the order of the specified linkage group(s) reversed.
}
\description{
This function reverses the order of a genetic map in a \code{mappoly} object.
It is typically used to align the MDS order with the genomic order.
}
\details{
The function modifies the order of markers in the specified linkage group(s) of
the genetic map. It reverses the positions of markers along with associated
data such as recombination fractions and haplotype probabilities, ensuring
that the genetic map's orientation is consistent with genomic data.
}
