% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{rf_filter}
\alias{rf_filter}
\title{Remove Markers Not Meeting LOD and Recombination Fraction Criteria}
\usage{
rf_filter(
  x,
  thresh.LOD.ph = 5,
  thresh.LOD.rf = 5,
  thresh.rf = 0.15,
  probs = c(0.05, 1),
  lg = NULL,
  type = c("mds", "genome"),
  diag.markers = NULL,
  mrk.order = NULL,
  diagnostic.plot = TRUE,
  breaks = 100
)
}
\arguments{
\item{x}{An object of class \code{mappoly2.data} or \code{mappoly2.sequence}.}

\item{thresh.LOD.ph}{LOD score threshold for linkage phase configuration.
Typically set to eliminate markers that are unlinked to the analyzed set (default = 5).}

\item{thresh.LOD.rf}{LOD score threshold for recombination fraction (default = 5).}

\item{thresh.rf}{Recombination fraction threshold (default = 0.15).}

\item{probs}{A numeric vector indicating the probability corresponding to the filtering quantiles (default = c(0.05, 1)).}

\item{lg}{A vector of linkage groups to be processed.
If NULL (default), all groups are considered.}

\item{type}{A character vector specifying the method for linkage group analysis.
Options are "mds" for multidimensional scaling and "genome" for genomic analysis.
This parameter only has an effect if \code{lg} is not NULL.}

\item{diag.markers}{A vector specifying a window of marker pairs to consider.
If NULL (default), all markers are considered.}

\item{mrk.order}{Marker order vector.
This parameter is only used if \code{diag.markers} is not NULL.}

\item{diagnostic.plot}{Logical; if \code{TRUE}, generates a diagnostic plot (default is TRUE).}

\item{breaks}{The number of cells for the histogram in the diagnostic plot (default = 100).}
}
\value{
A filtered object of the same class as the input (\code{mappoly2.data} or \code{mappoly2.sequence}).
}
\description{
This function removes markers from a \code{mappoly2.data} or \code{mappoly2.sequence} object
that do not meet specified LOD (logarithm of odds) and recombination fraction criteria.
It is designed to filter out markers that are unlikely to be linked or show strong
evidence of linkage across an entire linkage group, which might indicate false positives.
}
\details{
The function first checks the type of the input object and applies the
relevant filtering criteria based on LOD scores and recombination fractions.
It optionally produces a diagnostic plot to visualize the filtering process.
}
\examples{
\dontrun{
  # Assuming `my_data` is a valid mappoly2.data or mappoly2.sequence object
  filtered_data <- rf_filter(my_data, thresh.LOD.ph = 5, thresh.LOD.rf = 5)
}

}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu} with updates by Gabriel Gesteira, \email{gdesiqu@ncsu.edu}
}
