% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{rf_filter_per_group}
\alias{rf_filter_per_group}
\title{Filter Recombination Fractions for Specific Linkage Group}
\usage{
rf_filter_per_group(
  x,
  lg,
  type = c("mds", "genome"),
  thresh.LOD.ph = 5,
  thresh.LOD.rf = 5,
  thresh.rf = 0.15,
  probs = c(0.05, 1),
  diag.markers = NULL,
  diagnostic.plot = TRUE,
  breaks = 100
)
}
\arguments{
\item{x}{A \code{mappoly2.sequence} object.}

\item{lg}{Linkage group to filter.}

\item{type}{Method for linkage group analysis ("mds" or "genome").}

\item{thresh.LOD.ph}{LOD threshold for linkage phase (default = 5).}

\item{thresh.LOD.rf}{LOD threshold for recombination fraction (default = 5).}

\item{thresh.rf}{Recombination fraction threshold (default = 0.15).}

\item{probs}{Probability range for filtering (default = c(0.05, 1)).}

\item{diag.markers}{Window of marker pairs to consider (default = NULL).}

\item{diagnostic.plot}{Boolean to control the generation of a diagnostic plot (default = TRUE).}

\item{breaks}{Number of breaks for histogram in diagnostic plot (default = 100).}
}
\value{
A modified \code{mappoly2.sequence} object with filtered linkage group.
}
\description{
Internal function to filter recombination fractions for a specific linkage group in a \code{mappoly2.sequence} object.
}
\keyword{internal}
