% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{segreg_poly}
\alias{segreg_poly}
\title{Mendelian segregation}
\arguments{
\item{ploidy.p1}{ploidy level of parent 1}

\item{ploidy.p2}{ploidy level of parent 1}

\item{d.p1}{the dosage in parent 1}

\item{d.p2}{the dosage in parent 2}
}
\value{
a vector containing the expected segregation frequency for
genotypic classes.
}
\description{
Computes the Mendelian segregation frequencies given the ploidy level
of two parents and the dosage of the locus in both parents. It does
not consider double reduction.
}
\examples{
seg1 <- segreg_poly(ploidy_p1 = 6, ploidy_p2 = 6, d_p1 = 3, d_p2 = 3)
barplot(seg1)
seg2 <- segreg_poly(ploidy_p1 = 2, ploidy_p2 = 4, d_p1 = 2, d_p2 = 3)
barplot(seg2)

}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
