% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_phase}
\alias{sort_phase}
\title{Sort Phases Based on Log-Likelihood}
\usage{
sort_phase(ph, only.best = TRUE)
}
\arguments{
\item{ph}{A list of phases, each phase should have a \verb{$loglike} element.}

\item{only.best}{A boolean flag; if TRUE, only the best (highest log-likelihood) phase is returned.
If FALSE, all phases are returned in descending order of their log-likelihoods. Default is TRUE.}
}
\value{
If \code{only.best} is TRUE, returns the phase with the highest log-likelihood.
If FALSE, returns all phases sorted by log-likelihood in descending order.
}
\description{
This function sorts a list of phases based on their log-likelihood values.
It can return either the best phase (highest log-likelihood) or all phases in a sorted order.
}
\keyword{internal}
