% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.mappoly2.data}
\alias{subset.mappoly2.data}
\title{Data subset}
\usage{
\method{subset}{mappoly2.data}(
  x,
  type = c("marker", "individual"),
  perc = 0.1,
  n = NULL,
  select.mrk = NULL,
  select.ind = NULL,
  seed = NULL
)
}
\arguments{
\item{x}{an object  of class \code{mappoly2.data}}

\item{type}{should sample individuals or markers?}

\item{n}{number of individuals or markers to be sampled}

\item{percentage}{if \code{n == NULL}, the percentage of individuals or markers to be sampled}

\item{selected.ind}{a vector containing the name of the individuals to select. Only has effect
if \code{type = "individual"}, \code{n = NULL} and \code{percentage = NULL}}

\item{selected.mrk}{a vector containing the name of the markers to select. Only has effect
if \code{type = "marker"}, \code{n = NULL} and \code{percentage = NULL}}
}
\value{
an object  of class \code{mappoly2.data}
}
\description{
Data subset
}
