% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_data.R
\name{subset_data}
\alias{subset_data}
\title{Subsetting mappoly2 data set}
\usage{
subset_data(input.data, select.ind = NULL, select.mrk = NULL)
}
\arguments{
\item{x}{an object  of class \code{mappoly.data}}

\item{selected.ind}{a vector containing the name of the individuals to select.}

\item{selected.mrk}{a vector containing the name of the markers to select.}
}
\value{
an object  of class \code{mappoly.data}
}
\description{
Subsetting mappoly2 data set
}
\keyword{internal}
