% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{table_to_mappoly}
\alias{table_to_mappoly}
\title{Conversion of data.frame to mappoly.data}
\usage{
table_to_mappoly(
  dat,
  ploidy.p1,
  ploidy.p2,
  name.p1 = NULL,
  name.p2 = NULL,
  filter.non.conforming = TRUE,
  filter.redundant = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Conversion of data.frame to mappoly.data
}
\keyword{internal}
