% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_edits.R
\name{test_one_marker}
\alias{test_one_marker}
\title{Test All Possible Phases for a Given Marker in a Sequence}
\usage{
test_one_marker(
  x,
  mrk,
  lg = 1,
  type = c("mds", "genome", "custom"),
  parent = c("p1p2", "p1", "p2"),
  thresh.LOD.ph = 5,
  thresh.LOD.rf = 5,
  thresh.rf = 0.5,
  max.phases = 5,
  verbose = TRUE,
  tol = 0.001
)
}
\arguments{
\item{x}{A \code{mappoly2.sequence} object.}

\item{mrk}{The name of the marker to be tested.}

\item{lg}{The linkage group to which the marker belongs.}

\item{type}{The type of sequence ('mds', 'genome', or 'custom').}

\item{parent}{Specifies the parent phase ('p1p2', 'p1', or 'p2').}

\item{thresh.LOD.ph}{LOD threshold for linkage phase (default = 5).}

\item{thresh.LOD.rf}{LOD threshold for recombination fraction (default = 5).}

\item{thresh.rf}{Recombination fraction threshold (default = 0.5).}

\item{max.phases}{The maximum number of phase configurations to consider (default = 5).}

\item{verbose}{Logical; if \code{TRUE}, function prints messages during execution (default is \code{TRUE}).}

\item{tol}{Tolerance level for phase estimation (default = 10e-4).}
}
\value{
A list containing the best phasing results, including log-likelihood scores, recombination fractions, and phase information for each parent.
}
\description{
This internal function tests all possible phase configurations for a specified marker within a linkage group of a \code{mappoly2.sequence} object. It is designed to evaluate how well a marker fits into an existing genetic map by exploring different phasing scenarios.
}
\details{
The function operates as follows:
\enumerate{
\item Validates the input object and the specified marker.
\item Retrieves necessary information such as recombination fractions and dosage vectors for both parents.
\item Performs two-point phasing for each parent to determine the most likely phase configurations.
\item For each possible phase combination of the two parents, it estimates the Hidden Markov Model (HMM) map parameters.
\item Compares different phasing scenarios based on the log-likelihood of the HMM model fit.
\item The function returns a list of the best phase configurations based on the highest log-likelihood scores.
}
}
\keyword{internal}
