% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{text_col}
\alias{text_col}
\title{Changes the color of text in the R console, with a specific
focus on compatibility with RStudio's dark and light themes.}
\usage{
text_col(x, col = c("-", "red", "blue"))
}
\description{
Changes the color of text in the R console, with a specific
focus on compatibility with RStudio's dark and light themes.
}
