% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotypes.R
\name{calc_haplotypes}
\alias{calc_haplotypes}
\title{Conditional probabilities computation using Hidden Markov Models}
\usage{
calc_haplotypes(
  x,
  phase.conf = c("best", "all"),
  verbose = FALSE,
  tol = 0.001,
  compute.both.parents = TRUE
)
}
\arguments{
\item{void}{internal function}
}
\description{
Conditional probabilities computation using Hidden Markov Models
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
\keyword{internal}
