% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_rf.R
\name{est_pairwise_rf}
\alias{est_pairwise_rf}
\title{Pairwise Two-Point Analysis}
\usage{
est_pairwise_rf(
  input.seq,
  ncpus = 1L,
  mrk.pairs = NULL,
  verbose = TRUE,
  tol = .Machine$double.eps^0.25,
  ll = FALSE
)
}
\arguments{
\item{input.seq}{An object of class \code{mappoly2.sequence}}

\item{ncpus}{Number of parallel processes (cores) to use (default = 1)}

\item{mrk.pairs}{A 2xN matrix containing N pairs of markers to analyze. If \code{NULL} (default), all pairs are considered}

\item{verbose}{If \code{TRUE} (default), displays information about the analysis}

\item{tol}{Desired accuracy. See \code{optimize()} for details}

\item{ll}{If \code{TRUE}, returns log-likelihood instead of LOD scores (for internal use)}
}
\value{
An object of class \code{mappoly2.twopt} which is a list containing the following components:
\item{input.seq}{The input sequence; an object of class \code{mappoly2.sequence} with the raw data}
\item{pairwise}{A list where each element is a data frame. Row names have the format x-y, where x and y represent the
number of homolos sharing the same allelic variant in parents P1 and P2, respectively (see Mollinari and Garcia, 2019 for notation).
The first column contains the LOD Score in relation to the most likely linkage phase configuration. The second column shows the estimated
recombination fraction for each configuration, and the third column contains the LOD Score comparing the likelihood under no linkage (r = 0.5)
with the estimated recombination fraction (evidence of linkage).}
}
\description{
Performs a pairwise two-point analysis for all marker pairs in a given sequence.
The function estimates the recombination fraction for all possible linkage phase
configurations and their respective LOD Scores.
}
