% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_individuals}
\alias{filter_individuals}
\title{Filter out individuals}
\usage{
filter_individuals(x, ind.to.remove = NULL, inter = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{The name of the input object (class \code{mappoly.data}).}

\item{ind.to.remove}{A character vector containing the names of the individuals
to be removed. If \code{NULL}, the function opens an interactive graphic to
allow for individual selection.}

\item{inter}{If \code{TRUE}, the function expects user input to proceed with
filtering.}

\item{verbose}{If \code{TRUE} (default), the function shows the list of filtered
out individuals.}
}
\description{
This function removes individuals from the input dataset, either by specifying
them manually or by using interactive kinship analysis.
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
