% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_redundant}
\alias{filter_redundant}
\title{Filter Out Redundant Markers}
\usage{
filter_redundant(x, plot = TRUE)
}
\arguments{
\item{x}{An object of class \code{mappoly2.data}. It should have a component named \code{geno.dose}, which is a matrix containing the dosage for each marker (rows) for each individual (columns).}
}
\value{
The modified \code{mappoly2.data} object with redundant markers removed. The return value includes:
\itemize{
\item The original data with redundant markers removed.
\item The \code{redundant} component of the object updated to reflect changes. This component is a data frame listing the markers that were kept (\code{kept}) and the corresponding markers that were removed due to redundancy (\code{removed}).
}
}
\description{
Filters out redundant markers from a \code{mappoly2.data} object by identifying and removing duplicated genetic markers based on their genotypic information. Optionally, generates a plot showing the number of unique and redundant markers.
}
\details{
This function identifies redundant markers by checking for duplicates in the \code{geno.dose} matrix. If redundant markers are found, they are removed, and the object's \code{redundant} data frame is updated accordingly. If \code{plot = TRUE}, it generates a plot showing the number of markers before and after redundancy filtering.
}
\examples{
filtered_data <- filter_redundant(potato_dataset)

}
