% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_redundant}
\alias{filter_redundant}
\title{Filter out redundant markers}
\usage{
filter_redundant(input.seq)
}
\arguments{
\item{input.seq}{an object of class \code{mappoly.sequence}}
}
\value{
An object of class \code{mappoly2.seq}
}
\description{
Filter out markers with identical dosage information for all individuals.
}
