% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{genome_order}
\alias{genome_order}
\title{Get the genomic position of markers in a sequence}
\usage{
genome_order(x, mrk.names, verbose = TRUE)
}
\arguments{
\item{x}{an object of the class mappoly.geno.ord}

\item{verbose}{if \code{TRUE} (default), the current progress is shown; if
\code{FALSE}, no output is produced}

\item{...}{currently ignored}
}
\description{
This functions gets the genomic position of markers in a sequence and
return an ordered data frame with the name and position of each marker
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
