% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group}
\alias{group}
\title{Assign Markers to Linkage Groups}
\usage{
group(
  input.data = NULL,
  use.genome.only = FALSE,
  expected.groups = NULL,
  inter = TRUE,
  comp.mat = FALSE,
  LODweight = FALSE
)
}
\arguments{
\item{expected.groups}{The number of expected linkage groups (e.g. chromosomes)
for the species, if known.}

\item{inter}{A logical indicating whether to plot a dendrogram highlighting the
expected groups before continuing. Default is \code{TRUE}.}

\item{comp.mat}{A logical indicating whether to display a comparison between the
reference-based and linkage-based groupings if chromosome information is available.
Default is \code{FALSE}.}

\item{LODweight}{A logical indicating whether the clustering should be weighted by
the square of the LOD score. Default is \code{FALSE}.}

\item{input.mat}{A \code{mappoly2.rf.matrix} object containing the
recombination fraction and LOD score matrices.}

\item{input.seq}{A \code{mappoly2.sequence} object containing the markers
to be grouped. If \code{NULL} (default), all markers in \code{input.mat} are used.}
}
\value{
An object of class \code{mappoly2.group}, which is a list containing:
\itemize{
\item{hc.snp}{A list with information related to the UPGMA grouping method.}
\item{expected.groups}{The number of expected linkage groups.}
\item{groups.snp}{The assigned groups for each marker.}
\item{seq.vs.grouped.snp}{A comparison between the genomic group information
(when available) and the groups provided by \code{group}.}
}
}
\description{
This function identifies linkage groups of markers by utilizing the results
from two-point (pairwise) analysis.
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}#'
}
