% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{mapping}
\alias{mapping}
\title{Multi-Locus Map Estimation for mappoly2 Data Objects}
\usage{
mapping(
  x,
  lg = NULL,
  type = c("mds", "genome", "custom"),
  parent = c("p1p2", "p1", "p2"),
  recompute.from.pairwise = FALSE,
  phase.conf = "all",
  rf = NULL,
  error = 0,
  ncpus = 1,
  verbose = TRUE,
  tol = 0.001,
  ret_H0 = FALSE
)
}
\arguments{
\item{x}{An object of class \code{mappoly2.sequence}, representing the genetic map data.}

\item{lg}{Optional vector specifying linkage group indices to be processed. If NULL,
all linkage groups in the data object are considered.}

\item{type}{A character vector indicating the type of mapping to perform. Options
include "mds" (multi-dimensional scaling), "genome", and "custom".
Default is c("mds", "genome", "custom").}

\item{parent}{A character vector specifying the parent or parents to be considered
in the mapping process. Options are "p1p2" (both parents), "p1" (first parent),
and "p2" (second parent). Default is c("p1p2", "p1", "p2").}

\item{recompute.from.pairwise}{A logical value indicating whether to recompute the
map from pairwise data.}

\item{phase.conf}{A configuration parameter for phase determination.}

\item{rf}{Recombination fraction, used in the mapping calculations.}

\item{error}{Error tolerance in the mapping process.}

\item{ncpus}{Integer specifying the number of CPU cores for parallel processing.
Default is 1 (sequential processing).}

\item{verbose}{Logical value; if TRUE, detailed progress information is printed during
processing. Default is TRUE.}

\item{tol}{Tolerance level for the mapping algorithm.}

\item{ret_H0}{Logical; if TRUE, hypothesis testing results are returned.}
}
\value{
Returns an updated \code{mappoly2.sequence} data object with mapped linkage groups.
}
\description{
This function performs multi-locus map estimation on mappoly2 data objects, supporting
both sequential and parallel processing. It is designed to efficiently handle large
datasets, making it suitable for complex genetic mapping tasks.
}
\details{
The function processes each specified linkage group, performing phase determination
and map estimation based on the provided parameters. It can utilize parallel
processing to enhance performance on large datasets.
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
