% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{mapping}
\alias{mapping}
\title{Multipoint analysis using Hidden Markov Models}
\usage{
mapping(
  input.seq,
  phase.conf = "all",
  rf = NULL,
  error = 0,
  verbose = FALSE,
  tol = 0.001
)
}
\arguments{
\item{void}{internal function}
}
\description{
Multipoint analysis using Hidden Markov Models
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
\keyword{internal}
