% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{merge_single_parent_maps}
\alias{merge_single_parent_maps}
\title{This function merges two genetic maps built from markers that are exclusively
informative in isolated parents, facilitating unified analysis and visualization
of distinct genetic data.}
\usage{
merge_single_parent_maps(
  x,
  lg = NULL,
  type = c("mds", "genome", "custom"),
  hmm.reconstruction = TRUE,
  rf = NULL,
  error = 0,
  ncpus = 1,
  verbose = TRUE,
  tol = 0.001,
  ret_H0 = FALSE
)
}
\arguments{
\item{void}{internal function}
}
\description{
This function merges two genetic maps built from markers that are exclusively
informative in isolated parents, facilitating unified analysis and visualization
of distinct genetic data.
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
