% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasing.R
\name{pairwise_phasing}
\alias{pairwise_phasing}
\title{Phasing based on pairwise recombination fraction estimation}
\usage{
pairwise_phasing(
  x,
  lg = NULL,
  type = c("mds", "genome"),
  thresh.LOD.ph = 5,
  thresh.LOD.rf = 5,
  thresh.rf = 0.5,
  max.search.expansion.p1 = 10,
  max.search.expansion.p2 = max.search.expansion.p1,
  verbose = TRUE
)
}
\arguments{
\item{void}{internal function}
}
\description{
Phasing based on pairwise recombination fraction estimation
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
