% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{prepare_map}
\alias{prepare_map}
\title{prepare maps for plot}
\usage{
prepare_map(input.seq)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
prepare maps for plot
}
\keyword{internal}
