% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv.R
\name{read_geno_csv}
\alias{read_geno_csv}
\alias{print.mappoly2.data}
\alias{plot.mappoly2.data}
\title{Data Input in CSV format}
\usage{
read_geno_csv(
  file.in,
  ploidy.p1,
  ploidy.p2 = ploidy.p1,
  name.p1 = NULL,
  name.p2 = NULL,
  filter.non.conforming = TRUE,
  filter.redundant = TRUE,
  verbose = TRUE
)

\method{print}{mappoly2.data}(x, ...)

\method{plot}{mappoly2.data}(x, thresh.line = NULL, ...)
}
\arguments{
\item{file.in}{a character string with the name of (or full path to)
the input file containing the data to be read}

\item{ploidy.p1}{ploidy level of parent 1}

\item{ploidy.p2}{ploidy level of parent 1}

\item{name.p1}{name of parent 1}

\item{name.p2}{name of parent 2}

\item{filter.non.conforming}{if \code{TRUE} (default), data points with
unexpected genotypes (i.e. double reduction) are converted to 'NA'.
See the \code{\link[mappoly]{segreg_poly}} function for information on
expected classes and their respective frequencies.}

\item{filter.redundant}{logical. If \code{TRUE} (default), removes redundant
markers during map construction, keeping them annotated to export to the
final map.}

\item{verbose}{if \code{TRUE} (default), shows the current progress; if
\code{FALSE}, no output is produced}

\item{x}{an object of the class \code{mappoly.sequence}}

\item{...}{currently ignored}

\item{thresh.line}{position of a threshold line for p values of the segregation test (default = \code{0.05/n.mrk})}
}
\value{
An object of class \code{mappoly.data} which contains a list with
the following components:

\item{ploidy.p1}{ploidy level of the first parent}
\item{ploidy.p2}{ploidy level of the second parent}
\item{n.ind}{number individuals}
\item{n.mrk}{total number of markers}
\item{ind.names}{names or identifiers of the individuals}
\item{mrk.names}{names or identifiers of the genetic markers}
\item{name.p1}{names or identifiers of the first parent}
\item{name.p2}{names or identifiers of the second parent}
\item{dosage.p1}{the dosage for the first parent}
\item{dosage.p2}{the dosage for the second parent}
\item{chrom}{chromosome numbers for all markers}
\item{genome.pos}{physical positions on the genome for the genetic markers}
\item{seq.ref}{the reference DNA sequence data for the genetic markers}
\item{seq.alt}{the alternate DNA sequence data for the genetic markers}
\item{all.mrk.depth}{represents the depth of coverage for all genetic markers.
NULL when using csv input files}
\item{geno.dose}{a matrix containing the dosage for each marker (rows) for
each individual (columns).}
\item{kept}{a list of all non-redundant markers if \code{filter.redundant = TRUE}}
\item{filter.correspondence}{a list of all non-redundant markers and its equivalence
to the redundant ones if \code{filter.redundant = TRUE}}
}
\description{
Reads a comma-separated values (CSV) data file containing genetic marker data.
This function returns an object of class \code{mappoly.data}.
}
\details{
The CSV file should have rows representing markers, with the first row used
as a header. The first seven columns should contain the marker names, the
dosage in parents 1 and 2, the chromosome information (i.e. chromosome,
scaffold, contig, etc), the position of the marker within the sequence, the
alternate and reference alleles, if available.
The remaining columns should contain the dosage of the full-sib population.
For a tetraploid example of such a file, see the \code{Examples} section.
}
\examples{
\donttest{
tempfl <- list.files(system.file('extdata', package = 'mappoly2'),
                     full.names = TRUE)
SolCAP.dose <- read_geno_csv(file.in  = tempfl,
                                        ploidy.p1 = 4,
                                        name.p1 = "Atlantic",
                                        name.p2 = "B1829-5")
print(SolCAP.dose, detailed = TRUE)
plot(SolCAP.dose)
}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu} and
Gabriel Gesteira, \email{gdesiqu@ncsu.edu}
}
