% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.mappoly2.data}
\alias{subset.mappoly2.data}
\title{Subset a mappoly2.data Object}
\usage{
\method{subset}{mappoly2.data}(
  x,
  type = c("marker", "individual"),
  perc = 0.1,
  n = NULL,
  select.mrk = NULL,
  select.ind = NULL,
  seed = NULL,
  filter.non.conforming = TRUE,
  filter.redundant = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{mappoly2.data} object from which the subset will be created.}

\item{type}{Character string specifying whether to subset "marker" or "individual". Defaults to "marker".}

\item{perc}{Numeric value indicating the proportion of individuals or markers to sample if \code{n} is \code{NULL}. Defaults to 0.1.}

\item{n}{Integer specifying the number of individuals or markers to sample. If \code{NULL}, \code{perc} must be specified.}

\item{select.mrk}{Character vector of marker names to select. Used only if \code{type = "marker"}, and both \code{n} and \code{perc} are \code{NULL}.}

\item{select.ind}{Character vector of individual names to select. Used only if \code{type = "individual"}, and both \code{n} and \code{perc} are \code{NULL}.}

\item{seed}{Integer or \code{NULL}; sets the seed for random number generation for reproducible sampling.}

\item{filter.non.conforming}{Logical; if \code{TRUE} (default), data points with unexpected genotypes (e.g., double reduction) are converted to \code{NA}.}

\item{filter.redundant}{Logical; if \code{TRUE} (default), removes redundant markers during map construction, keeping them annotated for export in the final map.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A \code{mappoly2.data} object containing the selected subset of individuals or markers.
}
\description{
Creates a subset of individuals or markers from a \code{mappoly2.data} object based on specified criteria.
}
\details{
This function allows for flexible subsetting of \code{mappoly2.data} objects based on the number or proportion of individuals or markers. It supports random sampling and specific selection based on provided lists. Additional filtering options are available to handle non-conforming data and redundant markers.
}
\examples{
\donttest{
# Subset 20\% of markers
subset_data <- subset.mappoly2.data(mappoly_data, type = "marker", perc = 0.2)

# Subset 50 individuals
subset_data <- subset.mappoly2.data(mappoly_data, type = "individual", n = 50)

# Select specific markers
subset_data <- subset.mappoly2.data(mappoly_data, select.mrk = c("marker1", "marker2"))
}

}
