% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.mappoly2.data}
\alias{subset.mappoly2.data}
\title{Subset mappoly2.data Object}
\usage{
\method{subset}{mappoly2.data}(
  x,
  type = c("marker", "individual"),
  perc = 0.1,
  n = NULL,
  select.mrk = NULL,
  select.ind = NULL,
  seed = NULL,
  filter.non.conforming = TRUE,
  filter.redundant = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{mappoly2.data} object from which the subset is to be selected.}

\item{type}{Character string, either "marker" or "individual", indicating whether
markers or individuals should be subset. Default is "marker".}

\item{perc}{Numeric; the proportion of individuals or markers to be sampled.
This is used only if \code{n} is NULL. Default is 0.1.}

\item{n}{Integer; the number of individuals or markers to be sampled.
If NULL, \code{perc} must be specified.}

\item{select.mrk}{Character vector containing the names of the markers to select.
Effective only if \code{type = "marker"}, and both \code{n} and \code{perc} are NULL.}

\item{select.ind}{Character vector containing the names of the individuals to select.
Effective only if \code{type = "individual"}, and both \code{n} and \code{perc} are NULL.}

\item{seed}{Integer or NULL; sets the seed for random number generation for reproducible sampling.}

\item{filter.non.conforming}{Logical; if TRUE (default), data points with unexpected genotypes
(e.g., double reduction) are converted to 'NA'.}

\item{filter.redundant}{Logical; if TRUE (default), removes redundant markers during map construction,
keeping them annotated to export to the final map.}

\item{...}{Additional arguments, currently not used.}
}
\value{
A \code{mappoly2.data} object that contains the selected subset of individuals or markers.
}
\description{
This function creates a subset of either individuals or markers from an object
of the \code{mappoly2.data} class, based on specified criteria.
}
\details{
The function allows for flexible subsetting of \code{mappoly2.data} objects based on
the number or proportion of individuals or markers. It supports random sampling
and specific selection based on provided lists. Additional filtering options are
available to handle non-conforming data and redundant markers.
}
