% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.mappoly2.data}
\alias{subset.mappoly2.data}
\title{Subsets mappoly2.data Object}
\usage{
\method{subset}{mappoly2.data}(
  x,
  type = c("marker", "individual"),
  perc = 0.1,
  n = NULL,
  select.mrk = NULL,
  select.ind = NULL,
  seed = NULL,
  filter.non.conforming = TRUE,
  filter.redundant = TRUE
)
}
\arguments{
\item{x}{A \code{mappoly2.data} object from which the subset is to be selected.}

\item{type}{A character string, either "marker" or "individual", indicating whether markers or individuals should be subset.}

\item{n}{An integer specifying the number of individuals or markers to be sampled. If NULL, \code{percentage} must be specified.}

\item{seed}{a single value, interpreted as an integer, or \code{NULL}.
See the \code{\link[base]{set.seed}} function for information.}

\item{filter.non.conforming}{If \code{TRUE} (default), data points with
unexpected genotypes (e.g., double reduction) are converted to 'NA'.
See the \code{\link[mappoly]{segreg_poly}} function for information on
expected classes and their respective frequencies.}

\item{filter.redundant}{Logical. If \code{TRUE} (default), removes redundant
markers during map construction, keeping them annotated to export to the
final map.}

\item{percentage}{A numeric value between 0 and 1 indicating the proportion of individuals or markers to be sampled.
This parameter is considered only if \code{n} is set to NULL.}

\item{selected.ind}{A character vector containing the names of the individuals to select.
This parameter takes effect only if \code{type = "individual"}, and both \code{n} and \code{percentage} are NULL.}

\item{selected.mrk}{A character vector containing the names of the markers to select.
This parameter takes effect only if \code{type = "marker"}, and both \code{n} and \code{percentage} are NULL.}
}
\value{
A \code{mappoly2.data} object that contains the selected subset of individuals or markers.
}
\description{
This function selects a subset of either individuals or markers from an object of the \code{mappoly2.data} class.
}
