% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_edits.R
\name{test_marker}
\alias{test_marker}
\title{For a given marker, test all possible phases.}
\usage{
test_marker(input.seq, mrk, input.mat = NULL, tol = 0.001, verbose = TRUE)
}
\arguments{
\item{void}{internal function}
}
\description{
For a given marker, test all possible phases.
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
